% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.table.cmapComparison}
\alias{as.table.cmapComparison}
\title{Cross Tabulation and Table Creation}
\usage{
\method{as.table}{cmapComparison}(x, ..., clean = TRUE)
}
\arguments{
\item{x}{\code{cmapComparison} object}

\item{...}{Extra parameters passed to \code{table}}

\item{clean}{Boolean: only show certain columns (to avoid redundancy)?}
}
\value{
Complete table with metadata based on a \code{cmapComparison} object
}
\description{
Cross Tabulation and Table Creation
}
