% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{as.table.targetingDrugs}
\alias{as.table.targetingDrugs}
\title{Cross Tabulation and Table Creation}
\usage{
\method{as.table}{targetingDrugs}(x, ..., clean = TRUE)
}
\arguments{
\item{x}{\code{targetingDrugs} object}

\item{...}{Extra parameters passed to \code{table}}

\item{clean}{Boolean: only show certain columns (to avoid redundancy)?}
}
\value{
Complete table with metadata based on a \code{targetingDrugs} object
}
\description{
Cross Tabulation and Table Creation
}
\seealso{
Other functions related with the prediction of targeting drugs: 
\code{\link{listExpressionDrugSensitivityAssociation}()},
\code{\link{loadExpressionDrugSensitivityAssociation}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{predictTargetingDrugs}()}
}
\concept{functions related with the prediction of targeting drugs}
