% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{Calculate statistics}
\usage{
calculateStatistics(perturbations, cellLine, gene, coef)
}
\arguments{
\item{coef}{}
}
\value{
List of statistics
}
\description{
Calculate statistics
}
