% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapR_subset.R
\name{checkColnames}
\alias{checkColnames}
\title{Check whether \code{test_names} are columns in the \code{\link{data.frame}}}
\source{
https://github.com/cmap/cmapR
}
\usage{
checkColnames(test_names, df, throw_error = TRUE)
}
\arguments{
\item{test_names}{a vector of column names to test}

\item{df}{the \code{\link{data.frame}} to test against}

\item{throw_error}{boolean indicating whether to throw an error if
any \code{test_names} are not found in \code{df}}
}
\value{
boolean indicating whether or not all \code{test_names} are
  columns of \code{df}
}
\description{
Check whether \code{test_names} are columns in the \code{\link{data.frame}}
}
\keyword{internal}
