% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{chunkVector}
\alias{chunkVector}
\title{Divide columns into chunks to be loaded into memory at a time}
\usage{
chunkVector(x, nElems)
}
\description{
Divide columns into chunks to be loaded into memory at a time
}
\keyword{internal}
