% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapMetadata}
\alias{cmapMetadata}
\title{CMap metadata}
\description{
CMap metadata obtained by running the following code:

\preformatted{
cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
cmapMetadata <- filterCMapMetadata(cmapMetadata, cellLine = "HEPG2",
                                   timepoint = "2 h")
}
}
\keyword{internal}
