% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapPerturbationsKD}
\alias{cmapPerturbationsKD}
\title{CMap perturbations sample for knockdown experiments}
\description{
CMap perturbations sample for knockdown experiments obtained by running the
following code:

\preformatted{
# Code for loading CMap gene KD HepG2 data
cellLine <- "HepG2"
cmapMetadataKD <- filterCMapMetadata(
  "cmapMetadata.txt", cellLine=cellLine,
  perturbationType="Consensus signature from shRNAs targeting the same gene")

cmapPerturbationsKD <- loadCMapPerturbations(
  cmapMetadataKD, "cmapZscores.gctx", "cmapGeneInfo.txt")

# Select only some perturbations (to reduce file size)
data("diffExprStat")

compareKD <- compareAgainstCMap(diffExprStat, cmapPerturbationsKD)

filter <- c(head(order(compareKD$spearman_coef_rank)),
            tail(order(compareKD$spearman_coef_rank)),
            head(order(compareKD$pearson_coef_rank)),
            tail(order(compareKD$pearson_coef_rank)))
filter <- unique(compareKD[[1]][filter])
cmapPerturbationsKD <- cmapPerturbationsKD[ , filter]
}
}
\keyword{internal}
