% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapPerturbationsKnockdown}
\alias{cmapPerturbationsKnockdown}
\title{CMap perturbations sample for knockdown experiments}
\description{
CMap perturbations sample for knockdown experiments obtained by running the
following code:

\preformatted{
# Code for loading CMap gene KD HepG2 data
cellLine <- "HepG2"
cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
cmapMetadataKnockdown <- filterCMapMetadata(
  cmapMetadata, cellLine=cellLine,
  perturbationType="Consensus signature from shRNAs targeting the same gene")
cmapZscores  <- loadCMapData("cmapZscores.gctx", "zscores",
                             cmapMetadataKnockdown$sig_id)
cmapGeneInfo <- loadCMapData("cmapGeneInfo.txt", "geneInfo")

cmapPerturbationsKnockdown <- loadCMapPerturbations(
  cmapMetadataKnockdown, cmapZscores, cmapGeneInfo)

# Select only some perturbations (to reduce file size)
data("diffExprStat")

compareKnockdown <- list()
compareKnockdown$spearman <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="spearman")
compareKnockdown$pearson <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="pearson")
compareKnockdown$gsea <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="gsea",
    geneSize=150)

genes  <- lapply(compareKnockdown, "[[", "genes")
filter <- c(unlist(lapply(genes, head)), unlist(lapply(genes, tail)))
filter <- unique(filter)
cmapPerturbationsKnockdown <- cmapPerturbationsKnockdown[ , filter]
}
}
\keyword{internal}
