% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapPerturbationsSmallMolecules}
\alias{cmapPerturbationsSmallMolecules}
\title{CMap perturbations sample for small molecules}
\description{
CMap perturbations sample for small molecules obtained by running the
following code:

\preformatted{
cellLine <- c("HepG2", "HUH7")
cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
cmapMetadataSmallMolecules <- filterCMapMetadata(
    cmapMetadata, cellLine=cellLine, timepoint="24 h", dosage="5 µM",
    perturbationType="Compound")
cmapZscores  <- loadCMapData("cmapZscores.gctx", "zscores",
                             cmapMetadataSmallMolecules$sig_id)
cmapGeneInfo <- loadCMapData("cmapGeneInfo.txt")

cmapPerturbationsSmallMolecules <- loadCMapPerturbations(
    cmapMetadataSmallMolecules, cmapZscores, cmapGeneInfo)
}
}
\keyword{internal}
