% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseDuplicatedRows}
\alias{collapseDuplicatedRows}
\title{Collapse duplicated rows based on a column of a data frame}
\usage{
collapseDuplicatedRows(df, column)
}
\arguments{
\item{df}{Data frame}

\item{column}{Character: name of the column with elements to indicate which
rows to collapse}
}
\value{
Data frame with duplicated rows collapsed
}
\description{
Collapse duplicated rows based on a column of a data frame
}
\keyword{internal}
