% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compareAgainstL1000}
\alias{compareAgainstL1000}
\title{Compare against L1000 datasets}
\usage{
compareAgainstL1000(diffExprGenes, perturbations, cellLines, gene,
  method = c("spearman", "pearson", "gsea"), geneSize = 150)
}
\arguments{
\item{geneSize}{Number: top and bottom differentially expressed genes to use
for gene set enrichment (GSE); if \code{method} is not \code{gsea}, this
argument does nothing}
}
\description{
Compare against L1000 datasets
}
\examples{
#
compareAgainstL1000(diffExprGenes, perturbations, cellLines, gene)
}
