% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareAgainstReference}
\alias{compareAgainstReference}
\title{Compare multiple methods and rank reference accordingly}
\usage{
compareAgainstReference(
  diffExprGenes,
  reference,
  method = c("spearman", "pearson", "gsea"),
  geneSize = 150,
  cellLines = NULL,
  cellLineMean = "auto",
  rankByAscending = TRUE,
  rankPerCellLine = FALSE
)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
whose names are gene identifiers and respective values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{reference}{Data matrix or \code{perturbationChanges} object (CMap
perturbations; read \code{\link{prepareCMapPerturbations}})}

\item{method}{Character: comparison methods to run (\code{spearman},
\code{pearson} or \code{gsea}); multiple methods can be selected}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes for gene set enrichment (only used if \code{method = gsea})}

\item{cellLines}{Integer: number of unique cell lines}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default) the mean score will be
added if more than one cell line is available}

\item{rankByAscending}{Boolean: rank values based on their ascending (TRUE)
or descending (FALSE) order?}

\item{rankPerCellLine}{Boolean: when ranking results, also rank them based on
individual cell lines instead of only focusing on the mean score across
cell lines; if \code{cellLineMean = FALSE}, individual cell line conditions
are always ranked}
}
\value{
List of data frame containing the results per methods of comparison
}
\description{
Compare multiple methods and rank reference accordingly
}
