% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareAgainstReferencePerMethod}
\alias{compareAgainstReferencePerMethod}
\title{Compare single method}
\usage{
compareAgainstReferencePerMethod(
  method,
  input,
  reference,
  geneSize = 150,
  cellLines = NULL,
  cellLineMean = "auto",
  rankPerCellLine = FALSE
)
}
\arguments{
\item{method}{Character: comparison methods to run (\code{spearman},
\code{pearson} or \code{gsea}); multiple methods can be selected}

\item{input}{\code{Named numeric vector} of differentially expressed genes
whose names are gene identifiers and respective values are a statistic that
represents significance and magnitude of differentially expressed genes
(e.g. t-statistics); or \code{character} of gene symbols composing a gene
set that is tested for enrichment in reference data (only used if
\code{method = gsea})}

\item{reference}{Data matrix or \code{perturbationChanges} object (CMap
perturbations; see \code{\link{prepareCMapPerturbations}()})}

\item{geneSize}{Numeric: number of up-/down-regulated genes to use as gene
sets to test for enrichment in reference data (only used if
\code{method = gsea} and if \code{input} is a numeric vector)}

\item{cellLines}{Integer: number of unique cell lines}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default), the mean score will be
added when data for more than one cell line is available.}

\item{rankPerCellLine}{Boolean: rank results based on both individual cell
lines and mean scores across cell lines (\code{TRUE}) or based on mean
scores alone (\code{FALSE})? If \code{cellLineMean = FALSE}, individual
cell line conditions are always ranked.}

\item{rankByAscending}{Boolean: rank values based on their ascending
(\code{TRUE}) or descending (\code{FALSE}) order?}
}
\value{
Data frame containing the results per method of comparison
}
\description{
Compare single method
}
\keyword{internal}
