% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compareQuantile}
\alias{compareQuantile}
\title{Compare vector against its quantile}
\usage{
compareQuantile(vec, prob, lte = FALSE)
}
\arguments{
\item{vec}{Numeric vector}

\item{prob}{Numeric: probability value between [0,1] to produce sample
quantiles}

\item{lte}{Boolean: check if values are <= quantile? If \code{FALSE}, checks
if values are >= quantile}
}
\value{
Boolean vector regarding compared elements
}
\description{
Check which elements of the vector are lower/greater than or equal to the
quantile of a given vector.
}
\keyword{internal}
