% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{compareSingleMethod}
\alias{compareSingleMethod}
\title{Compare single method}
\usage{
compareSingleMethod(method, diffExprGenes = diffExprGenes,
  perturbations = perturbations, geneSize = geneSize,
  cellLineMean = cellLineMean)
}
\arguments{
\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea}; multiple methods may be selected at once)}

\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{cmapPerturbations} object: CMap perturbations
(check \code{\link{loadCMapPerturbations}})}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes to use for gene set enrichment (GSE) (only used if
\code{method = gsea})}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default) the mean score will be
added if more than one cell line is available}
}
\description{
Compare single method
}
\keyword{internal}
