% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{compareWithCellLineProgress}
\alias{compareWithCellLineProgress}
\title{Compare with cell line (print progress)}
\usage{
compareWithCellLineProgress(cellLine, cellLines, FUN, method,
  perturbations, metadata, ...)
}
\arguments{
\item{metadata}{Data frame: perturbation metadata}
}
\description{
Compare with cell line (print progress)
}
\keyword{internal}
