% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertENSEMBLtoGeneSymbols}
\alias{convertENSEMBLtoGeneSymbols}
\title{Convert ENSEMBL gene identifiers to gene symbols}
\usage{
convertENSEMBLtoGeneSymbols(
  genes,
  dataset = "hsapiens_gene_ensembl",
  mart = "ensembl"
)
}
\arguments{
\item{genes}{Character: ENSEMBL gene identifiers}

\item{dataset}{Character: \code{biomaRt} dataset name}

\item{mart}{Character: \code{biomaRt} database name}
}
\value{
Named character vector where names are the input ENSEMBL gene
  identifiers and the values are the matching gene symbols
}
\description{
Convert ENSEMBL gene identifiers to gene symbols
}
