% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{correlateAgainstCMap}
\alias{correlateAgainstCMap}
\title{Correlate against CMap perturbations}
\usage{
correlateAgainstCMap(diffExprGenes, perturbations, method, cellLines,
  pAdjust = "BH")
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{perturbationChanges} object: CMap perturbations
(check \code{\link{prepareCMapPerturbations}})}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea}; multiple methods may be selected at once)}

\item{pAdjust}{Character: method to use for p-value adjustment}
}
\value{
Data frame with correlation results per perturbation
}
\description{
Correlate against CMap perturbations
}
\keyword{internal}
