% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{correlateAgainstReference}
\alias{correlateAgainstReference}
\title{Correlate against data columns}
\usage{
correlateAgainstReference(diffExprGenes, reference, method,
  pAdjust = "BH")
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
whose names are gene identifiers and respective values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea}; multiple methods may be selected at once)}

\item{pAdjust}{Character: method to use for p-value adjustment}
}
\value{
Data frame with correlation results per data column
}
\description{
Correlate against data columns
}
\keyword{internal}
