% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{correlatePerCellLine}
\alias{correlatePerCellLine}
\title{Correlate differential expression scores per cell line}
\usage{
correlatePerCellLine(cellLine, diffExprGenes, perturbations, filtered,
  method, pAdjustMethod = "BH")
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{cmapPerturbations} object: CMap perturbations
(check \code{\link{loadCMapPerturbations}})}

\item{filtered}{Character: perturbations to filter}

\item{method}{Character: correlation method}
}
\value{
Data frame with correlations statistics, p-value and q-value
}
\description{
Correlate differential expression scores per cell line
}
\keyword{internal}
