% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{diffExprStat}
\alias{diffExprStat}
\title{Differential expression's t-statistics sample}
\description{
Differential expression's t-statistics sample obtained by running the
following code:

\preformatted{
data("counts")

# Perform differential gene expression analysis
diffExpr <- performDifferentialExpression(counts)

# Get t-statistics of differential expression with respective gene names
diffExprStat <- diffExpr$t
names(diffExprStat) <- diffExpr$Gene_symbol
}
}
\keyword{internal}
