% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.cmapPerturbations}
\alias{dim.cmapPerturbations}
\title{Dimensions of a \code{cmapPerturbations} object}
\usage{
\method{dim}{cmapPerturbations}(x)
}
\arguments{
\item{x}{\code{cmapPerturbations} object}
}
\value{
Dimensions of a \code{cmapPerturbations} object
}
\description{
Dimensions of a \code{cmapPerturbations} object
}
