% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimnames.cmapPerturbations}
\alias{dimnames.cmapPerturbations}
\title{Dimnames of a \code{cmapPerturbations} object}
\usage{
\method{dimnames}{cmapPerturbations}(x)
}
\arguments{
\item{x}{\code{cmapPerturbations} object}
}
\value{
Retrieve dimnames of a \code{cmapPerturbations} object
}
\description{
Dimnames of a \code{cmapPerturbations} object
}
