% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{dimnames.expressionDrugSensitivityAssociation}
\alias{dimnames.expressionDrugSensitivityAssociation}
\alias{dim.expressionDrugSensitivityAssociation}
\alias{[.expressionDrugSensitivityAssociation}
\title{Operations on \code{expressionDrugSensitivityAssociation} objects}
\usage{
\method{dimnames}{expressionDrugSensitivityAssociation}(x)

\method{dim}{expressionDrugSensitivityAssociation}(x)

\method{[}{expressionDrugSensitivityAssociation}(x, i, j, drop = FALSE, ...)
}
\arguments{
\item{x}{An \code{expressionDrugSensitivityAssociation} object}

\item{i, j}{Character or numeric indexes specifying elements to extract}

\item{drop}{Boolean: coerce result to the lowest possible dimension?}

\item{...}{Extra arguments given to other methods}
}
\value{
Subset, dimension or dimension names
}
\description{
Operations on \code{expressionDrugSensitivityAssociation} objects
}
