% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{dimnames.perturbationChanges}
\alias{dimnames.perturbationChanges}
\title{Dimnames of a \code{perturbationChanges} object}
\usage{
\method{dimnames}{perturbationChanges}(x)
}
\arguments{
\item{x}{\code{perturbationChanges} object}
}
\value{
Retrieve dimnames of a \code{perturbationChanges} object
}
\description{
Dimnames of a \code{perturbationChanges} object
}
\seealso{
Other functions related with the ranking of CMap perturbations: 
\code{\link{[.perturbationChanges}()},
\code{\link{as.table.similarPerturbations}()},
\code{\link{dim.perturbationChanges}()},
\code{\link{filterCMapMetadata}()},
\code{\link{getCMapConditions}()},
\code{\link{getCMapPerturbationTypes}()},
\code{\link{loadCMapData}()},
\code{\link{loadCMapZscores}()},
\code{\link{parseCMapID}()},
\code{\link{plot.perturbationChanges}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{prepareCMapPerturbations}()},
\code{\link{print.similarPerturbations}()},
\code{\link{rankSimilarPerturbations}()}
}
\concept{functions related with the ranking of CMap perturbations}
