% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{.plotBubbles}
\alias{.plotBubbles}
\title{Plot packed bubbles}
\usage{
.plotBubbles(data, title, colour = "orange")
}
\arguments{
\item{data}{Data to plot}

\item{title}{Character: plot title}

\item{colour}{Character: bubble colour}
}
\value{
\code{highchart} object
}
\description{
Plot packed bubbles
}
\keyword{internal}
