% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{downloadENCODEknockdownMetadata}
\alias{downloadENCODEknockdownMetadata}
\title{Load metadata for ENCODE knockdown experiments}
\usage{
downloadENCODEknockdownMetadata(cellLine = NULL, gene = NULL)
}
\description{
Load metadata for ENCODE knockdown experiments
}
