% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{downloadENCODEsamples}
\alias{downloadENCODEsamples}
\title{Load ENCODE samples}
\usage{
downloadENCODEsamples(metadata)
}
\arguments{
\item{metadata}{Character: ENCODE metadata}
}
\value{
List of loaded ENCODE samples
}
\description{
Samples are automatically downloaded and loaded if they are not found in the
current workspace.
}
\examples{
if (interactive()) {
  # Download ENCODE metadata for a specific cell line and gene
  cellLine <- "HepG2"
  gene <- "EIF4G1"
  ENCODEmetadata <- downloadENCODEknockdownMetadata(cellLine, gene)

  # Download samples based on filtered ENCODE metadata
  ENCODEsamples <- downloadENCODEsamples(ENCODEmetadata)
}
}
