% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadIfNotFound}
\alias{downloadIfNotFound}
\title{Download data if given file is not found}
\usage{
downloadIfNotFound(link, file, ask = FALSE, toExtract = NULL)
}
\arguments{
\item{link}{Character: link to download file}

\item{file}{Character: filepath}

\item{ask}{Boolean: ask to download file?}

\item{toExtract}{Character: files to extract (if \code{NULL}, extract all)}
}
\value{
Download file if file is not found
}
\description{
Download data if given file is not found
}
\keyword{internal}
