% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{downloadL1000data}
\alias{downloadL1000data}
\title{Load L1000 data}
\usage{
downloadL1000data(file, type = c("metadata", "geneInfo", "zscores",
  "compoundInfo"), zscoresId = NULL)
}
\arguments{
\item{file}{Character: path to file}

\item{type}{Character: type of data to load (\code{metadata},
\code{geneInfo}, \code{zscores} or \code{compoundInfo})}

\item{zscoresId}{Character: identifiers to partially load z-scores file
(for performance reasons)}
}
\value{
Metadata as a data table
}
\description{
Load L1000 data. If \code{file} does not exist, it will first be downloaded.
}
\note{
If \code{type = compoundInfo}, two files from
\strong{The Drug Repurposing Hub} will be downloaded containing information
about drugs and perturbations. The files will be named \code{file} with
\code{_drugs} and \code{_samples} before their extension, respectively.
}
\examples{
# Download L1000 metadata
l1000metadata <- downloadL1000data("l1000metadata.txt", "metadata")

# Download L1000 gene info
downloadL1000data("l1000geneInfo.txt", "geneInfo")

# Download L1000 zscores based on filtered metadata
l1000metadataKnockdown <- filterL1000metadata(
  l1000metadata, cellLine="HepG2",
  perturbationType="Consensus signature from shRNAs targeting the same gene")

if (interactive()) {
    downloadL1000data("l1000zscores.gctx.gz", "zscores",
                      l1000metadataKnockdown$sig_id)
}
}
