% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1000-package.r
\docType{data}
\name{ENCODEmetadata}
\alias{ENCODEmetadata}
\title{ENCODE metadata sample}
\description{
ENCODE metadata sample obtained by running the following code:

\preformatted{
gene <- "EIF4G1"
cellLine <- "HepG2"
ENCODEmetadata <- downloadENCODEknockdownMetadata(cellLine, gene)
}
}
