% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{ENCODEsamples}
\alias{ENCODEsamples}
\title{Sample of ENCODE samples}
\description{
Sample of ENCODE samples obtained by running the following code:

\preformatted{
data("ENCODEmetadata")

# Load samples based on filtered ENCODE metadata
ENCODEsamples <- loadENCODEsamples(ENCODEmetadata)[[1]]

# Get small subset of whole dataset
filter <- ENCODEsamples[[1]]$expected_count > 0
genes <- head(ENCODEsamples[[1]][filter, ])$gene_id
for (k in seq(ENCODEsamples)) {
    ENCODEsamples[[k]] <- ENCODEsamples[[k]][
        ENCODEsamples[[k]]$gene_id %in% genes, ]
}
}
}
\keyword{internal}
