% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{filterL1000metadata}
\alias{filterL1000metadata}
\title{Filter L1000 metadata}
\usage{
filterL1000metadata(metadata, cellLine = NULL, timepoint = NULL,
  dosage = NULL, perturbationType = NULL)
}
\arguments{
\item{metadata}{Data frame: metadata}

\item{cellLine}{Character: cell line (if \code{NULL}, all values are loaded)}

\item{timepoint}{Character: timepoint (if \code{NULL}, all values are loaded)}

\item{dosage}{Character: dosage (if \code{NULL}, all values are loaded)}

\item{perturbationType}{Character: type of perturbation (if \code{NULL}, all
perturbation types are loaded)}
}
\value{
Filtered L1000 metadata
}
\description{
Filter L1000 metadata
}
\examples{
data("l1000metadata")
# l1000metadata <- loadL1000data("l1000metadata.txt", "metadata")
filterL1000metadata(l1000metadata, cellLine="HEPG2", timepoint="2 h",
                    dosage="25 ng/mL")
}
