% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{getCMapConditions}
\alias{getCMapConditions}
\title{List available conditions in CMap datasets}
\usage{
getCMapConditions(metadata, cellLine = NULL, timepoint = NULL,
  dosage = NULL, perturbationType = NULL, control = FALSE)
}
\arguments{
\item{metadata}{Data frame (CMap metadata) or character (respective filepath)}

\item{cellLine}{Character: cell line (if \code{NULL}, all values are loaded)}

\item{timepoint}{Character: timepoint (if \code{NULL}, all values are loaded)}

\item{dosage}{Character: dosage (if \code{NULL}, all values are loaded)}

\item{perturbationType}{Character: type of perturbation (if \code{NULL}, all
perturbation types are loaded)}

\item{control}{Boolean: show controls for perturbation types?}
}
\value{
List of conditions in CMap datasets
}
\description{
Downloads metadata if not available
}
\examples{
data("cmapMetadata")
# cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
}
