% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{getCMapPerturbationTypes}
\alias{getCMapPerturbationTypes}
\title{Get perturbation types}
\usage{
getCMapPerturbationTypes(control = FALSE)
}
\arguments{
\item{control}{Boolean: return perturbation types used as control?}
}
\value{
Perturbation types and respective codes as used by CMap datasets
}
\description{
Get perturbation types
}
\examples{
getCMapPerturbationTypes()
}
\seealso{
Other functions related with the ranking of CMap perturbations: \code{\link{[.perturbationChanges}},
  \code{\link{as.table.similarPerturbations}},
  \code{\link{dim.perturbationChanges}},
  \code{\link{dimnames.perturbationChanges}},
  \code{\link{filterCMapMetadata}},
  \code{\link{getCMapConditions}},
  \code{\link{loadCMapData}},
  \code{\link{loadCMapZscores}}, \code{\link{parseCMapID}},
  \code{\link{plot.perturbationChanges}},
  \code{\link{plot.referenceComparison}},
  \code{\link{plotTargetingDrugsVSsimilarPerturbations}},
  \code{\link{prepareCMapPerturbations}},
  \code{\link{print.similarPerturbations}},
  \code{\link{rankSimilarPerturbations}}
}
\concept{functions related with the ranking of CMap perturbations}
