% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{getENCODEcontrols}
\alias{getENCODEcontrols}
\title{Get experiments files for a given control}
\usage{
getENCODEcontrols(control, table)
}
\arguments{
\item{control}{Character: control identifier}

\item{table}{Data frame}
}
\value{
Character vector with respective experiment identifiers
}
\description{
Get experiments files for a given control
}
