% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{getL1000conditions}
\alias{getL1000conditions}
\title{List available conditions in L1000 datasets}
\usage{
getL1000conditions(metadata)
}
\arguments{
\item{metadata}{frame: L1000 metadata}
}
\value{
List of conditions in L1000 datasets
}
\description{
Downloads metadata if not available
}
\examples{
data("l1000metadata")
# l1000metadata <- downloadL1000data("l1000metadata.txt", "metadata")
getL1000conditions(l1000metadata)
}
