% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{getL1000perturbationTypes}
\alias{getL1000perturbationTypes}
\title{Get perturbation types}
\usage{
getL1000perturbationTypes()
}
\value{
Perturbation types and respective codes as used by L1000 datasets
}
\description{
Get perturbation types
}
\examples{
getL1000perturbationTypes()
}
