% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{l1000metadata}
\alias{l1000metadata}
\title{L1000 metadata}
\description{
L1000 metadata obtained by running the following code:

\preformatted{
l1000metadata <- downloadL1000data("l1000metadata.txt", "metadata")
l1000metadata <- filterL1000metadata(l1000metadata, cellLine = "HEPG2",
                                     timepoint = "2 h")
}
}
\keyword{internal}
