% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1000-package.r
\docType{data}
\name{l1000perturbationsKnockdown}
\alias{l1000perturbationsKnockdown}
\title{L1000 perturbations sample for knockdown experiments}
\description{
L1000 perturbations sample for knockdown experiments obtained by running the
following code:

\preformatted{
# Code for loading CMap gene KD HepG2 data
l1000metadata <- downloadL1000data("l1000metadata.txt", "metadata")
l1000metadataKnockdown <- filterL1000metadata(
  l1000metadata, cellLine="HepG2",
  perturbationType="Consensus signature from shRNAs targeting the same gene")
l1000zscores  <- downloadL1000data("l1000zscores.gctx", "zscores",
                                   l1000metadataKnockdown$sig_id)
l1000geneInfo <- downloadL1000data("l1000geneInfo.txt", "geneInfo")

l1000perturbationsKnockdown <- loadL1000perturbations(
  l1000metadataKnockdown, l1000zscores, l1000geneInfo)

# Select only some perturbations (to reduce file size)
data("diffExprStat")

compareKnockdown <- list()
compareKnockdown$spearman <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="spearman")
compareKnockdown$pearson <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="pearson")
compareKnockdown$gsea <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="gsea",
    geneSize=150)

genes  <- lapply(compareKnockdown, "[[", "genes")
filter <- c(unlist(lapply(genes, head)), unlist(lapply(genes, tail)))
l1000perturbationsKnockdown <- l1000perturbationsKnockdown[ , filter]
}
}
