% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{l1000perturbationsSmallMolecules}
\alias{l1000perturbationsSmallMolecules}
\title{L1000 perturbations sample for small molecules}
\description{
L1000 perturbations sample for small molecules obtained by running the
following code:

\preformatted{
l1000metadataSmallMolecules <- filterL1000metadata(
    l1000metadata, cellLine="HepG2", timepoint="24 h",
    dosage="5 \\U00B5M", # \\U00B5 is the unicode code for the micro symbol
    perturbationType="Compound")
l1000zscores  <- downloadL1000data("l1000zscores.gctx", "zscores",
                                   l1000metadataSmallMolecules$sig_id)
l1000geneInfo <- downloadL1000data("l1000geneInfo.txt")

l1000perturbationsSmallMolecules <- loadL1000perturbations(
    l1000metadataSmallMolecules, l1000zscores, l1000geneInfo,
    sanitizeCompoundNames=TRUE)

# Select only some perturbations (to reduce file size)
data("diffExprStat")

compareSmallMolecule <- list()
compareSmallMolecule$spearman <- compareAgainstL1000(
    diffExprStat, l1000perturbationsSmallMolecules, cellLine,
    method="spearman")
compareSmallMolecule$pearson <- compareAgainstL1000(
    diffExprStat, l1000perturbationsSmallMolecules, cellLine,
    method="pearson")
compareSmallMolecule$gsea <- compareAgainstL1000(
    diffExprStat, l1000perturbationsSmallMolecules, cellLine, method="gsea",
    geneSize=150)

genes  <- lapply(compareSmallMolecule, "[[", "genes")
filter <- c(unlist(lapply(genes, head)), unlist(lapply(genes, tail)))
l1000perturbationsSmallMolecules <- l1000perturbationsSmallMolecules[
    , filter]
}
}
\keyword{internal}
