% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchDiffExprLoader}
\alias{launchDiffExprLoader}
\title{Load differential expression data via a visual interface}
\usage{
launchDiffExprLoader(
  cellLine = NULL,
  gene = NULL,
  file = "ENCODEmetadata.rds",
  path = "."
)
}
\arguments{
\item{cellLine}{Character: cell line}

\item{gene}{Character: target gene}

\item{file}{Character: RDS filepath with metadata (if file doesn't exist, it
will be created)}

\item{path}{Character: path where to download files}
}
\value{
Differential expression data
}
\description{
Currently only supports loading data from ENCODE knockdown experiments
}
\seealso{
Other visual interface functions: 
\code{\link{cTRAP}()},
\code{\link{launchCMapDataLoader}()},
\code{\link{launchDrugSetEnrichmentAnalyser}()},
\code{\link{launchMetadataViewer}()},
\code{\link{launchResultPlotter}()}
}
\concept{visual interface functions}
