% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchDrugSetEnrichmentAnalyser}
\alias{launchDrugSetEnrichmentAnalyser}
\title{View and plot results via a visual interface}
\usage{
launchDrugSetEnrichmentAnalyser(sets, ...)
}
\arguments{
\item{sets}{Named list of characters: named sets containing compound
identifiers (obtain drug sets by running \code{prepareDrugSets()})}

\item{...}{Objects}
}
\value{
Launches result viewer and plotter (returns \code{NULL})
}
\description{
View and plot results via a visual interface
}
\seealso{
Other visual interface functions: 
\code{\link{cTRAP}()},
\code{\link{launchCMapDataLoader}()},
\code{\link{launchDiffExprLoader}()},
\code{\link{launchMetadataViewer}()},
\code{\link{launchResultPlotter}()}
}
\concept{visual interface functions}
