% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchMetadataViewer}
\alias{launchMetadataViewer}
\title{View metadata via a visual interface}
\usage{
launchMetadataViewer(...)
}
\arguments{
\item{...}{Objects}
}
\value{
Metadata viewer (retunrs \code{NULL})
}
\description{
View metadata via a visual interface
}
\seealso{
Other visual interface functions: 
\code{\link{launchCMapDataLoader}()},
\code{\link{launchDiffExprLoader}()},
\code{\link{launchDrugSetEnrichmentAnalyser}()},
\code{\link{launchResultPlotter}()}
}
\concept{visual interface functions}
