% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchResultPlotter}
\alias{launchResultPlotter}
\title{View and plot results}
\usage{
launchResultPlotter(...)
}
\arguments{
\item{...}{Objects}
}
\value{
Launches result viewer and plotter
}
\description{
View and plot results
}
