% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{listExpressionDrugSensitivityAssociation}
\alias{listExpressionDrugSensitivityAssociation}
\title{List available gene expression and drug sensitivity correlation matrices}
\usage{
listExpressionDrugSensitivityAssociation(url = FALSE)
}
\arguments{
\item{url}{Boolean: return download link?}
}
\value{
Character vector of available gene expression and drug sensitivity
  correlation matrices
}
\description{
List available gene expression and drug sensitivity correlation matrices
}
\examples{
listExpressionDrugSensitivityAssociation()
}
\seealso{
Other functions related with the prediction of targeting drugs: 
\code{\link{loadExpressionDrugSensitivityAssociation}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{predictTargetingDrugs}()}
}
\concept{functions related with the prediction of targeting drugs}
