% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{loadCMapPerturbations}
\alias{loadCMapPerturbations}
\title{Load CMap perturbation data}
\usage{
loadCMapPerturbations(metadata, zscores, geneInfo, compoundInfo = NULL)
}
\arguments{
\item{metadata}{Data frame (CMap metadata) or character (respective filepath)}

\item{zscores}{Data frame (GCTX z-scores) or character (respective filepath)}

\item{geneInfo}{Data frame (CMap gene info) or character (respective
filepath)}

\item{compoundInfo}{Data frame (CMap compound info) or character (respective
filepath)}
}
\value{
Perturbation data from CMap as a data table
}
\description{
Load CMap perturbation data
}
\examples{
if (interactive()) {
  metadata <- loadCMapData("cmapMetadata.txt", "metadata")
  metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
  zscores  <- loadCMapData("cmapZscores.gctx", "zscores", metadata$sig_id)
  geneInfo <- loadCMapData("cmapGeneInfo.txt", "geneInfo")
  loadCMapPerturbations(metadata, zscores, geneInfo)
}
}
