% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadENCODEexperiment}
\alias{loadENCODEexperiment}
\title{Load an ENCODE experiment for a given gene}
\usage{
loadENCODEexperiment(data, gene)
}
\description{
Load an ENCODE experiment for a given gene
}
