% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{loadENCODEgeneExpression}
\alias{loadENCODEgeneExpression}
\title{Load an ENCODE gene expression data for a given gene}
\usage{
loadENCODEgeneExpression(metadata)
}
\arguments{
\item{metadata}{Character: ENCODE metadata}

\item{gene}{Character: gene}
}
\value{
Data frame containing gene read counts for which ENCODE knockdown
experiment of a given gene
}
\description{
Load an ENCODE gene expression data for a given gene
}
