% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{loadENCODEmetadata}
\alias{loadENCODEmetadata}
\title{Load ENCODE metadata}
\usage{
loadENCODEmetadata(file, cellLine = NULL, gene = NULL, lab = NULL,
  outputType = NULL)
}
\arguments{
\item{file}{Character: file path to ENCODE metadata file}

\item{cellLine}{Character: cell line}

\item{gene}{Character: target gene}

\item{lab}{Character: laboratory}

\item{outputType}{Character: output type}
}
\value{
ENCODE metadata loaded as a data table
}
\description{
Load ENCODE metadata
}
