% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{loadENCODEsample}
\alias{loadENCODEsample}
\title{Load ENCODE sample}
\usage{
loadENCODEsample(metadata, replicate, control = FALSE)
}
\arguments{
\item{metadata}{Data frame: ENCODE metadata}

\item{replicate}{Number: replicate}

\item{control}{Boolean: load control experiment?}
}
\description{
Load ENCODE sample
}
