% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{loadExpressionDrugSensitivityAssociation}
\alias{loadExpressionDrugSensitivityAssociation}
\title{Load gene expression and drug sensitivity correlation matrix}
\usage{
loadExpressionDrugSensitivityAssociation(source, file = NULL)
}
\arguments{
\item{source}{Character: source (\code{CTRP}, \code{GDSC} or \code{NCI60})}

\item{file}{Character: filepath to gene expression and drug sensitivity
association dataset (automatically downloaded if file does not exist)}
}
\value{
Correlation matrix between gene expression (rows) and drug
  sensitivity (columns)
}
\description{
Load gene expression and drug sensitivity correlation matrix
}
\examples{
loadExpressionDrugSensitivityAssociation("GDSC")
}
