% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadInternalData}
\alias{loadInternalData}
\title{Load internal data for use in vignettes and examples}
\usage{
loadInternalData(x)
}
\arguments{
\item{x}{Character: name of object to load}
}
\value{
Object
}
\description{
Load internal data for use in vignettes and examples
}
\examples{
loadInternalData("l1000perturbationsKnockdown")
}
