% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{loadL1000file}
\alias{loadL1000file}
\title{Load data from L1000 files}
\usage{
loadL1000file(cellLine, timepoint, folder)
}
\arguments{
\item{cellLine}{Character: cell line}

\item{timepoint}{Character: time point}

\item{folder}{Character: folder where L1000 files are stored}
}
\value{
L1000 file as a data table
}
\description{
Load data from L1000 files
}
