% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{loadL1000metadata}
\alias{loadL1000metadata}
\title{Load L1000 metadata}
\usage{
loadL1000metadata(l1000metadataFile)
}
\arguments{
\item{l1000metadataFile}{Character: filepath to L1000 metadata}
}
\value{
Metadata as a data table
}
\description{
The data will be downloaded if not available
}
