% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{loadL1000perturbations}
\alias{loadL1000perturbations}
\title{Load L1000 perturbation data}
\usage{
loadL1000perturbations(l1000metadataFile, l1000zscoresFile, l1000geneFile,
  cellLine = NULL, timepoint = NULL, dosage = NULL,
  perturbationType = NULL, sanitizeCompoundNames = FALSE)
}
\arguments{
\item{l1000metadataFile}{Character: filepath to L1000 metadata}

\item{l1000zscoresFile}{Character: path to GCTX z-scores file}

\item{l1000geneFile}{Character: path to L1000 gene file}

\item{cellLine}{Character: cell line (if \code{NULL}, all values are loaded)}

\item{timepoint}{Character: timepoint (if \code{NULL}, all values are loaded)}

\item{dosage}{Character: dosage (if \code{NULL}, all values are loaded)}

\item{perturbationType}{Character: type of perturbation (if \code{NULL}, all
perturbation types are loaded)}

\item{sanitizeCompoundNames}{Boolean: replace identifiers with compound names}
}
\value{
Perturbation data from L1000 as a data table
}
\description{
Load L1000 perturbation data
}
