% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{loadNCI60drugSensitivity}
\alias{loadNCI60drugSensitivity}
\title{Load CTRP data}
\usage{
loadNCI60drugSensitivity(file = "NCI60/drugSensitivity.xls")
}
\arguments{
\item{file}{Character: file path}
}
\value{
Data frame
}
\description{
If given paths direct to non-existing files, those files will be downloaded
}
\keyword{internal}
