% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapR_subset.R
\name{parse_gctx}
\alias{parse_gctx}
\title{Parse a GCTX file into the workspace as a GCT object}
\usage{
parse_gctx(fname, cid = NULL)
}
\arguments{
\item{fname}{path to the GCTX file on disk}

\item{cid}{either a vector of character or integer
column indices or a path to a grp file containing character
column indices. Only these indicies will be parsed from the
file.}
}
\description{
Parse a GCTX file into the workspace as a GCT object
}
\details{
\code{parse_gctx} also supports parsing of plain text
  GCT files, so this function can be used as a general GCT parser.
}
\examples{
gct_file <- system.file("extdata", "modzs_n272x978.gctx", package="cmapR")
(ds <- parse_gctx(gct_file))

# matrix only
(ds <- parse_gctx(gct_file, matrix_only=TRUE))

# only the first 10 rows and columns
(ds <- parse_gctx(gct_file, rid=1:10, cid=1:10))

}
\seealso{
Other GCTX parsing functions: \code{\link{fix.datatypes}},
  \code{\link{process_ids}}, \code{\link{read.gctx.ids}},
  \code{\link{read.gctx.meta}}
}
\concept{GCTX parsing functions}
