% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{performGSAperCellLine}
\alias{performGSAperCellLine}
\title{Perform gene set enrichment (GSA) per cell line}
\usage{
performGSAperCellLine(cellLine, perturbations, pathways)
}
\arguments{
\item{cellLine}{Character: cell line(s)}

\item{perturbations}{\code{cmapPerturbations} object: file with CMap loaded
perturbations (check \code{\link{loadCMapPerturbations}})}

\item{pathways}{List of gene sets to check.}
}
\value{
Data frame containing gene set enrichment analysis (GSEA) results per
cell line
}
\description{
Perform gene set enrichment (GSA) per cell line
}
\keyword{internal}
