% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{performGSEAagainstCMap}
\alias{performGSEAagainstCMap}
\title{Perform gene set enrichment (GSA) against CMap perturbations}
\usage{
performGSEAagainstCMap(diffExprGenes, perturbations, pathways, cellLines)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{cmapPerturbations} object: CMap perturbations
(check \code{\link{prepareCMapPerturbations}})}

\item{pathways}{List of gene sets to check.}
}
\value{
Data frame containing gene set enrichment analysis (GSEA) results per
perturbation
}
\description{
Perform gene set enrichment (GSA) against CMap perturbations
}
\keyword{internal}
