% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cmapComparison}
\alias{plot.cmapComparison}
\title{Plot CMap data comparison}
\usage{
\method{plot}{cmapComparison}(x, perturbation = NULL,
  perturbationID = NULL, genes = c("both", "top", "bottom"))
}
\arguments{
\item{x}{\code{cmapComparison} object}

\item{perturbationID}{Character: perturbation identifier}

\item{genes}{Character: when plotting gene set enrichment analysis (GSEA),
plot top genes (\code{genes = "top"}), bottom genes
(\code{genes = "bottom"}) or both (\code{genes = "both"})}
}
\value{
Plot illustrating the comparison with CMap data
}
\description{
Plot CMap data comparison
}
\examples{
data("cmapPerturbationsKnockdown")
cellLine <- "HepG2"
compareKnockdown <- list()

# Compare against CMap using Spearman correlation
compareKnockdown$spearman <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="spearman")

# Compare against CMap using Pearson correlation
compareKnockdown$pearson <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="pearson")

# Compare against CMap using gene set enrichment analysis (GSEA) with the
# top and bottom 150 genes as gene sets
compareKnockdown$gsea <- compareAgainstCMap(
    diffExprStat, cmapPerturbationsKnockdown, cellLine, method="gsea",
    geneSize=150)

EIF4G1knockdown <- grep("EIF4G1", compareKnockdown$gsea$genes, value=TRUE)
plot(compareKnockdown$spearman, EIF4G1knockdown)
plot(compareKnockdown$pearson, EIF4G1knockdown)
plot(compareKnockdown$gsea, EIF4G1knockdown)
}
