% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cmapPerturbations}
\alias{plot.cmapPerturbations}
\title{Plot CMap data comparison}
\usage{
\method{plot}{cmapPerturbations}(x, perturbation, diffExprGenes,
  method = c("spearman", "pearson", "gsea"), geneSize = 150,
  genes = c("both", "top", "bottom"))
}
\arguments{
\item{x}{\code{cmapPerturbations} object}

\item{perturbation}{Character (perturbation identifier) or a
\code{cmapComparison} table (from which the respective perturbation
identifiers are retrieved)}

\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{method}{Character: method to plot results (\code{spearman},
\code{pearson} or \code{gsea})}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes to use for gene set enrichment (GSE) (only used if
\code{method = gsea})}

\item{genes}{Character: when plotting gene set enrichment analysis (GSEA),
plot top genes (\code{genes = "top"}), bottom genes
(\code{genes = "bottom"}) or both (\code{genes = "both"}); only used if
\code{method = gsea}}
}
\description{
Plot CMap data comparison
}
\examples{
data("diffExprStat")
data("cmapPerturbationsKD")

# Compare against CMap using Spearman's correlation, Pearson's correlation
# and gene set enrichment analysis (GSEA) with the top and bottom 150 genes
# as gene sets
compareKD <- compareAgainstCMap(diffExprStat, cmapPerturbationsKD)

EIF4G1knockdown <- grep("EIF4G1", compareKD[[1]], value=TRUE)

plot(cmapPerturbationsKD, EIF4G1knockdown, diffExprStat, method="spearman")
plot(cmapPerturbationsKD, EIF4G1knockdown, diffExprStat, method="pearson")
plot(cmapPerturbationsKD, EIF4G1knockdown, diffExprStat, method="gsea")
}
