% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.L1000comparison}
\alias{plot.L1000comparison}
\title{Plot L1000 data comparison}
\usage{
\method{plot}{L1000comparison}(object, perturbationID, topGenes = TRUE)
}
\arguments{
\item{object}{\code{L1000 comparison} object}

\item{perturbationID}{Character: perturbation identifier}

\item{topGenes}{Boolean: plot top (\code{topGenes = TRUE}) or bottom genes
(\code{topGenes = FALSE}) in case of plotting gene set enrichment analysis
(GSEA)}

\item{cellLine}{Character: cell line}
}
\value{
Plot illustrating the comparison with L1000 data
}
\description{
Plot L1000 data comparison
}
