% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotESplot}
\alias{plotESplot}
\title{Render GSEA enrichment plot}
\usage{
plotESplot(enrichmentScore, gseaStat, compact = FALSE)
}
\value{
GSEA enrichment plot
}
\description{
Render GSEA enrichment plot
}
\keyword{internal}
