% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGSEAenrichment}
\alias{plotGSEAenrichment}
\title{Render GSEA enrichment plot}
\usage{
plotGSEAenrichment(enrichmentScore, gseaStat, titleSize = 14,
  axisTitleSize = 12, axisTextSize = 10, pointSize = 0.1)
}
\value{
GSEA enrichment plot
}
\description{
Render GSEA enrichment plot
}
\keyword{internal}
