% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRes}
\alias{plotRes}
\title{Plot perturbations according to a given column representation}
\usage{
plotRes(x, cellLine = NULL, n = 5, signifThreshold = 0.05,
  alpha = 0.3)
}
\description{
Plot perturbations according to a given column representation
}
\keyword{internal}
