% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSingleCorr}
\alias{plotSingleCorr}
\title{Render scatter plot to show a single relationship}
\usage{
plotSingleCorr(perturbation, label, diffExprGenes)
}
\description{
Render scatter plot to show a single relationship
}
\keyword{internal}
