% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTargetingDrugsVSsimilarPerturbations}
\alias{plotTargetingDrugsVSsimilarPerturbations}
\title{Plot similar perturbations against predicted targeting drugs}
\usage{
plotTargetingDrugsVSsimilarPerturbations(targetingDrugs,
  similarPerturbations, column, labelColumn = "pert_iname",
  quantileThreshold = 0.25, showAllScores = FALSE, ...)
}
\arguments{
\item{similarPerturbations}{\code{similarPerturbations} object}

\item{column}{Character: column to plot (must be available in both databases)}

\item{labelColumn}{Character: column in \code{similarPerturbations}, its
metadata or compound information to be used for labelling}

\item{showAllScores}{Boolean: showl all scores? If \code{FALSE}, only the
best score per compound will be plotted}

\item{x}{\code{targetingDrugs} object}
}
\value{
\code{ggplot2} plot
}
\description{
Plot similar perturbations against predicted targeting drugs
}
