% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{predictTargetingDrugs}
\alias{predictTargetingDrugs}
\title{Predict targeting drugs}
\usage{
predictTargetingDrugs(diffExprGenes, expressionDrugSensitivityCor,
  method = c("spearman", "pearson", "gsea"), geneSize = 150)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
whose names are gene identifiers and respective values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{expressionDrugSensitivityCor}{Matrix: correlation matrix of gene
expression (rows) and drug sensitivity (columns) across cell lines.
Pre-prepared gene expression and drug sensitivity associations are
available to download using
\code{\link{loadExpressionDrugSensitivityAssociation}}.}

\item{method}{Character: comparison methods to run (\code{spearman},
\code{pearson} or \code{gsea}); multiple methods can be selected}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes for gene set enrichment (only used if \code{method = gsea})}
}
\value{
Data table with correlation or GSEA results comparing differential
  expression values against gene expression and drug sensitivity associations
}
\description{
Identify compounds that may target the phenotype associated with a
user-provided differential expression profile by comparing such against a
correlation matrix of gene expression and drug sensitivity.
}
\section{GSEA score}{

Weighted connectivity scores (WTCS) are calculated when \code{method
  = "gsea"} (\url{https://clue.io/connectopedia/cmap_algorithms}).
}

\examples{
# Example of a differential expression profile
data("diffExprStat")

# Load expression and drug sensitivity association derived from GDSC data
gdsc <- loadExpressionDrugSensitivityAssociation("GDSC")

# Predict targeting drugs on a differential expression profile
predictTargetingDrugs(diffExprStat, gdsc)
}
