% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{prepareENCODEgeneExpression}
\alias{prepareENCODEgeneExpression}
\title{Load an ENCODE gene expression data}
\usage{
prepareENCODEgeneExpression(samples)
}
\arguments{
\item{samples}{List of loaded ENCODE samples}
}
\value{
Data frame containing gene read counts
}
\description{
Load an ENCODE gene expression data
}
\examples{
data("ENCODEsamples")
## Load ENCODE metadata for a specific cell line and gene
# cellLine <- "HepG2"
# gene <- "EIF4G1"
# ENCODEmetadata <- loadENCODEknockdownMetadata(cellLine, gene)

## Load samples based on filtered ENCODE metadata
# ENCODEsamples <- loadENCODEsamples(ENCODEmetadata)

prepareENCODEgeneExpression(ENCODEsamples)
}
