% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{prepareExpressionDrugSensitivityAssociation}
\alias{prepareExpressionDrugSensitivityAssociation}
\title{Prepare gene expression and drug sensitivity correlation matrix}
\usage{
prepareExpressionDrugSensitivityAssociation(
  dataset = c("GDSC 7", "CTRP 2.1", "NCI60"),
  method = "spearman"
)
}
\arguments{
\item{dataset}{Character: dataset to use (\code{CTRP}, \code{GDSC} or
\code{NCI60})}

\item{method}{Character: correlation method to use between gene expression
and drug sensitivity}
}
\value{
Correlation matrix between gene expression and drug sensitivity
}
\description{
Prepare gene expression and drug sensitivity correlation matrix
}
\details{
If path directs to non-existing files, data will be downloaded.
}
\keyword{internal}
