% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{prepareGSEApathways}
\alias{prepareGSEApathways}
\title{Prepare GSEA pathways}
\usage{
prepareGSEApathways(diffExprGenes, geneSize)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
whose names are gene identifiers and respective values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes for gene set enrichment (only used if \code{method = gsea})}
}
\value{
List of top and bottom differentially expressed genes
}
\description{
Prepare GSEA pathways
}
\keyword{internal}
