% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{prepareSetsCompoundInfo}
\alias{prepareSetsCompoundInfo}
\title{Get drug sets' compound info}
\usage{
prepareSetsCompoundInfo(sets)
}
\description{
Get drug sets' compound info
}
\keyword{internal}
