% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{prepareStatsCompoundInfo}
\alias{prepareStatsCompoundInfo}
\title{Prepare stats' compound information}
\usage{
prepareStatsCompoundInfo(stats)
}
\description{
Prepare stats' compound information
}
\keyword{internal}
