% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{processPertsPerChunk}
\alias{processPertsPerChunk}
\title{Process perturbation changes per chunk}
\usage{
processPertsPerChunk(perturbations, FUN, ..., chunkSize = 10000)
}
\arguments{
\item{perturbations}{\code{perturbationChanges} object: CMap perturbations
(check \code{\link{prepareCMapPerturbations}})}

\item{FUN}{Function: function to run for each chunk}

\item{...}{Arguments passed to \code{FUN}}

\item{chunkSize}{Integer: number of perturbations to load on-demand (a higher
value increases RAM usage, but also time-wise performance)}
}
\value{
Results of running \code{FUN}
}
\description{
Perturbations will be processed per chunk if argument \code{perturbations} is
a file path instead of a data matrix. Otherwise, the data will be processed
in full (i.e., as a single chunk).
}
\details{
Loading a chunk of 10000 CMap pertubations requires ~1GB of RAM.
}
\keyword{internal}
