% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapR_subset.R
\name{read.gctx.meta}
\alias{read.gctx.meta}
\title{Parse row or column metadata from GCTX files}
\source{
https://github.com/cmap/cmapR
}
\usage{
read.gctx.meta(gctx_path, dimension = "row", ids = NULL,
  set_annot_rownames = TRUE)
}
\arguments{
\item{gctx_path}{the path to the GCTX file}

\item{dimension}{which metadata to read (row or column)}

\item{ids}{a character vector of a subset of row/column ids
for which to read the metadata}

\item{set_annot_rownames}{a boolean indicating whether to set the
\code{rownames} addtribute of the returned \code{data.frame} to
the corresponding row/column ids.}
}
\value{
a \code{data.frame} of metadata
}
\description{
Parse row or column metadata from GCTX files
}
\examples{
gct_file <- system.file("extdata", "modzs_n272x978.gctx", package="cmapR")
# row meta
row_meta <- read.gctx.meta(gct_file)
str(row_meta)
# column meta
col_meta <- read.gctx.meta(gct_file, dimension="column")
str(col_meta)
# now for only the first 10 ids
col_meta_first10 <- read.gctx.meta(gct_file, dimension="column",
                                   ids=col_meta$id[1:10])
str(col_meta_first10)

}
\seealso{
Other GCTX parsing functions: \code{\link{fix.datatypes}},
  \code{\link{process_ids}}, \code{\link{read.gctx.ids}}
}
\concept{GCTX parsing functions}
