% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapR_subset.R
\name{readGctxIds}
\alias{readGctxIds}
\title{Read GCTX row or column ids}
\source{
\url{https://github.com/cmap/cmapR}
}
\usage{
readGctxIds(gctx_path, dimension = "row")
}
\arguments{
\item{gctx_path}{path to the GCTX file}

\item{dimension}{which ids to read (row or column)}
}
\value{
a character vector of row or column ids from the provided file
}
\description{
Read GCTX row or column ids
}
\seealso{
Other GCTX parsing functions: 
\code{\link{fix.datatypes}()},
\code{\link{processIds}()},
\code{\link{readGctxMeta}()}
}
\concept{GCTX parsing functions}
\keyword{internal}
