% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stripStr}
\alias{stripStr}
\title{Strip non-alpha-numeric characters from a string}
\usage{
stripStr(str)
}
\arguments{
\item{str}{Character}
}
\value{
Character without non-alphanumeric values
}
\description{
Strip non-alpha-numeric characters from a string
}
\keyword{internal}
