% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.cmapPerturbations}
\alias{[.cmapPerturbations}
\title{Subset an \code{cmapPerturbations} object}
\usage{
\method{[}{cmapPerturbations}(x, ...)
}
\arguments{
\item{x}{\code{cmapPerturbations} object}

\item{...}{Extra parameters passed to \code{`[`}}
}
\value{
\code{cmapPerturbations} object with subset data
}
\description{
Subset an \code{cmapPerturbations} object
}
