% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.l1000perturbations}
\alias{[.l1000perturbations}
\title{Subset an \code{l1000perturbations} object}
\usage{
\method{[}{l1000perturbations}(x, ...)
}
\arguments{
\item{x}{\code{l1000perturbations} object}

\item{...}{Extra parameters passed to \code{`[`}}
}
\value{
\code{l1000perturbations} object with subset data
}
\description{
Subset an \code{l1000perturbations} object
}
