% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{[.perturbationChanges}
\alias{[.perturbationChanges}
\title{Subset a \code{perturbationChanges} object}
\usage{
\method{[}{perturbationChanges}(x, i, j, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{perturbationChanges} object}

\item{i, j}{Character or numeric indexes specifying elements to extract}

\item{drop}{Boolean: coerce result to the lowest possible dimension?}

\item{...}{Extra parameters passed to \code{`[`}}
}
\value{
\code{perturbationChanges} object with subset data
}
\description{
Subset a \code{perturbationChanges} object
}
\seealso{
Other functions related with the ranking of CMap perturbations: \code{\link{as.table.similarPerturbations}},
  \code{\link{dim.perturbationChanges}},
  \code{\link{dimnames.perturbationChanges}},
  \code{\link{filterCMapMetadata}},
  \code{\link{getCMapConditions}},
  \code{\link{getCMapPerturbationTypes}},
  \code{\link{loadCMapData}},
  \code{\link{loadCMapZscores}}, \code{\link{parseCMapID}},
  \code{\link{plot.perturbationChanges}},
  \code{\link{plot.referenceComparison}},
  \code{\link{plotTargetingDrugsVSsimilarPerturbations}},
  \code{\link{prepareCMapPerturbations}},
  \code{\link{print.similarPerturbations}},
  \code{\link{rankSimilarPerturbations}}
}
\concept{functions related with the ranking of CMap perturbations}
