% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetData}
\alias{subsetData}
\title{Subset data by rows and/or columns}
\usage{
subsetData(x, i, j, rowAttr, colAttr, nargs, ...)
}
\value{
Subset data
}
\description{
Subset data by rows and/or columns
}
\keyword{internal}
