% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetDim}
\alias{subsetDim}
\title{Subset rows or columns based on a given index}
\usage{
subsetDim(k, dims, nargs, areCols = TRUE)
}
\value{
Subset rows/columns
}
\description{
Subset rows or columns based on a given index
}
\keyword{internal}
