% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapR_subset.R
\name{subsetToIds}
\alias{subsetToIds}
\title{Do a robust \code{\link{data.frame}} subset to a set of ids}
\source{
https://github.com/cmap/cmapR
}
\usage{
subsetToIds(df, ids)
}
\arguments{
\item{df}{\code{\link{data.frame}} to subset}

\item{ids}{the ids to subset to}
}
\value{
a subset version of \code{df}
}
\description{
Do a robust \code{\link{data.frame}} subset to a set of ids
}
\keyword{internal}
