% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{as.table.similarPerturbations}
\alias{as.table.similarPerturbations}
\title{Cross Tabulation and Table Creation}
\usage{
\method{as.table}{similarPerturbations}(x, ..., clean = TRUE)
}
\arguments{
\item{x}{\code{similarPerturbations} object}

\item{...}{Extra parameters passed to \code{table}}

\item{clean}{Boolean: only show certain columns (to avoid redundancy)?}
}
\value{
Complete table with metadata based on a \code{similarPerturbations}
  object
}
\description{
Cross Tabulation and Table Creation
}
\seealso{
Other functions related with the ranking of CMap perturbations: \code{\link{[.perturbationChanges}},
  \code{\link{dim.perturbationChanges}},
  \code{\link{dimnames.perturbationChanges}},
  \code{\link{filterCMapMetadata}},
  \code{\link{getCMapConditions}},
  \code{\link{getCMapPerturbationTypes}},
  \code{\link{loadCMapData}},
  \code{\link{loadCMapZscores}}, \code{\link{parseCMapID}},
  \code{\link{plot.perturbationChanges}},
  \code{\link{plot.referenceComparison}},
  \code{\link{plotTargetingDrugsVSsimilarPerturbations}},
  \code{\link{prepareCMapPerturbations}},
  \code{\link{print.similarPerturbations}},
  \code{\link{rankSimilarPerturbations}}
}
\concept{functions related with the ranking of CMap perturbations}
