% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{calculateCellLineMean}
\alias{calculateCellLineMean}
\title{Calculate cell line mean}
\usage{
calculateCellLineMean(data, cellLine, rankCellLinePerturbations)
}
\arguments{
\item{data}{Data table: comparison against CMap data}

\item{cellLine}{Character: perturbation identifiers as names and respective
cell lines as values}

\item{rankCellLinePerturbations}{Boolean: when ranking results, also rank
perturbations regarding individual cell lines instead of the mean
perturbation score alone; if \code{cellLineMean = FALSE}, individual cell
line perturbations are always ranked}
}
\value{
A list with two items:
\describe{
\item{\code{data}}{input \code{data} with extra rows containing cell line
  average scores (if calculated)}
\item{\code{cellLineInfo}}{data table with cell line information}
}
}
\description{
Calculate cell line mean
}
\keyword{internal}
