% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{chunkVector}
\alias{chunkVector}
\title{Assign vector elements into chunks}
\usage{
chunkVector(x, nElems)
}
\arguments{
\item{x}{Vector of elements}

\item{nElems}{Numeric: number of chunks}
}
\value{
List of chunks with the original vector elements divided
}
\description{
Assign vector elements into chunks
}
\keyword{internal}
