% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapPerturbationsKnockdown}
\alias{cmapPerturbationsKnockdown}
\title{CMap perturbations sample for knockdown experiments}
\description{
CMap perturbations sample for knockdown experiments obtained by running the
following code:

\preformatted{
# Code for loading CMap gene KD HepG2 data
cellLine <- "HepG2"
cmapMetadataKnockdown <- filterCMapMetadata(
  "cmapMetadata.txt", cellLine=cellLine,
  perturbationType="Consensus signature from shRNAs targeting the same gene")

cmapPerturbationsKnockdown <- loadCMapPerturbations(
  cmapMetadataKnockdown, "cmapZscores.gctx", "cmapGeneInfo.txt")

# Select only some perturbations (to reduce file size)
data("diffExprStat")

compareKnockdown <- compareAgainstCMap(diffExprStat,
                                       cmapPerturbationsKnockdown)

filter <- c(head(order(compareKnockdown$spearman_coef_rank)),
            tail(order(compareKnockdown$spearman_coef_rank)),
            head(order(compareKnockdown$pearson_coef_rank)),
            tail(order(compareKnockdown$pearson_coef_rank)))
filter <- c(unlist(lapply(genes, head)), unlist(lapply(genes, tail)))
filter <- unique(filter)
cmapPerturbationsKnockdown <- cmapPerturbationsKnockdown[ , filter]
}
}
\keyword{internal}
