% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTRAP-package.r
\docType{data}
\name{cmapPerturbationsSmallMolecules}
\alias{cmapPerturbationsSmallMolecules}
\title{CMap perturbations sample for small molecules}
\description{
CMap perturbations sample for small molecules obtained by running the
following code:

\preformatted{
cellLine <- "HepG2"
cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
cmapMetadataSmallMolecules <- filterCMapMetadata(
    cmapMetadata, cellLine=cellLine, timepoint="24 h",
    dosage="5 \\U00B5M", # \\U00B5 is the unicode code for the micro symbol
    perturbationType="Compound")
cmapZscores  <- loadCMapData("cmapZscores.gctx", "zscores",
                             cmapMetadataSmallMolecules$sig_id)
cmapGeneInfo <- loadCMapData("cmapGeneInfo.txt")

cmapPerturbationsSmallMolecules <- loadCMapPerturbations(
    cmapMetadataSmallMolecules, cmapZscores, cmapGeneInfo)
}
}
\keyword{internal}
