% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{compareAgainstCMap}
\alias{compareAgainstCMap}
\title{Compare against CMap datasets}
\usage{
compareAgainstCMap(diffExprGenes, perturbations, method = c("spearman",
  "pearson", "gsea"), geneSize = 150, pAdjustMethod = "BH",
  cellLineMean = "auto")
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{cmapPerturbations} object: file with CMap loaded
perturbations (check \code{\link{loadCMapPerturbations}})}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea})}

\item{geneSize}{Number: top and bottom differentially expressed genes to use
for gene set enrichment (GSE) (only used if \code{method} is \code{gsea})}

\item{pAdjustMethod}{Character: method for p-value adjustment (for more
details, see \code{\link{p.adjust.methods}}; only used if \code{method} is
\code{spearman} or \code{pearson})}

\item{cellLineMean}{Boolean: add a column with the mean across cell lines? If
\code{"auto"} (default) the mean will be added if more than one cell line is
available}

\item{cellLine}{Character: cell line(s)}
}
\value{
Data table with correlation or GSEA results comparing differential
gene expression values with those associated with CMap perturbations
}
\description{
Weighted connectivity scores (WTCS) are calculated when \code{method} is
\code{gsea}. For more information on WTCS, read
\url{https://clue.io/connectopedia/cmap_algorithms}.
}
\details{
Order results according to the correlation coefficient (if
\code{method} is \code{spearman} or \code{pearson}) or the weighted
connectivity score (WTCS) score (if \code{method} is \code{gsea}) of the mean
across cell lines (if \code{cellLineMean} is \code{TRUE}; otherwise results
are ordered based on the first cell line alone).
}
\examples{
cellLine <- "HepG2"
data("cmapPerturbationsSmallMolecules")
perturbations <- cmapPerturbationsSmallMolecules
data("diffExprStat")

# Compare against CMap using Spearman correlation
compareAgainstCMap(diffExprStat, perturbations, cellLine,
                   method="spearman")

# Compare against CMap using Pearson correlation
compareAgainstCMap(diffExprStat, perturbations, cellLine,
                   method="pearson")

# Compare against CMap using gene set enrichment analysis (GSEA)
compareAgainstCMap(diffExprStat, perturbations, cellLine, method="gsea")
}
