% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{compareAgainstL1000}
\alias{compareAgainstL1000}
\title{Compare against L1000 datasets}
\usage{
compareAgainstL1000(diffExprGenes, perturbations, cellLine,
  method = c("spearman", "pearson", "gsea"), geneSize = 150,
  pAdjustMethod = "BH", cellLineMean = length(cellLine) > 1)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{l1000perturbations} object: file with L1000 loaded
perturbations (check \code{\link{loadL1000perturbations}})}

\item{cellLine}{Character: cell line(s)}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea})}

\item{geneSize}{Number: top and bottom differentially expressed genes to use
for gene set enrichment (GSE) (only used if \code{method} is \code{gsea})}

\item{pAdjustMethod}{Character: method for p-value adjustment (for more
details, see \code{\link{p.adjust.methods}}; only used if \code{method} is
\code{spearman} or \code{pearson})}

\item{cellLineMean}{Boolean: add a column with the mean across cell lines}
}
\value{
Data table with correlation or GSEA results comparing differential
gene expression values with those associated with L1000 perturbations
}
\description{
Weighted connectivity scores (WCTS) are calculated when \code{method} is
\code{gsea}. For more information on WCTS, read
\url{https://clue.io/connectopedia/cmap_algorithms}.
}
\details{
Order results according to the correlation coefficient (if
\code{method} is \code{spearman} or \code{pearson}) or the weighted
connectivity score (WTCS) score (if \code{method} is \code{gsea}) of the mean
across cell lines (if \code{cellLineMean} is \code{TRUE}; otherwise results
are ordered based on the first cell line alone).
}
\examples{
cellLine <- "HepG2"
data("l1000perturbationsSmallMolecules")
perturbations <- l1000perturbationsSmallMolecules
data("diffExprStat")

# Compare against L1000 using Spearman correlation
compareAgainstL1000(diffExprStat, perturbations, cellLine,
                    method="spearman")

# Compare against L1000 using Pearson correlation
compareAgainstL1000(diffExprStat, perturbations, cellLine,
                    method="pearson")

# Compare against L1000 using gene set enrichment analysis (GSEA)
compareAgainstL1000(diffExprStat, perturbations, cellLine, method="gsea")
}
