% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{compareAgainstL1000}
\alias{compareAgainstL1000}
\title{Compare against L1000 datasets}
\usage{
compareAgainstL1000(diffExprGenes, perturbations, cellLine,
  method = c("spearman", "pearson", "gsea"), geneSize = 150)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{L1000perturbations} object: file with L1000 loaded
perturbations (check \code{\link{loadL1000perturbations}})}

\item{cellLine}{Character: cell line(s)}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea})}

\item{geneSize}{Number: top and bottom differentially expressed genes to use
for gene set enrichment (GSE); if \code{method} is not \code{gsea}, this
argument does nothing}
}
\value{
Data table with correlation or GSEA results comparing differential
gene expression values with those associated with L1000 perturbations
}
\description{
Compare against L1000 datasets
}
\examples{
cellLine <- "HepG2"
perturbations <- l1000:::l1000perturbationsSmallMolecules
diffExprGenes <- l1000:::diffExprStat

# Compare against L1000 using Spearman correlation
compareAgainstL1000(diffExprGenes, perturbations, cellLine,
                    method="spearman")

# Compare against L1000 using Pearson correlation
compareAgainstL1000(diffExprGenes, perturbations, cellLine,
                    method="pearson")

# Compare against L1000 using gene set enrichment analysis (GSEA)
compareAgainstL1000(diffExprGenes, perturbations, cellLine, method="gsea")
}
