% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{compareAgainstL1000}
\alias{compareAgainstL1000}
\title{Compare against L1000 datasets}
\usage{
compareAgainstL1000(diffExprGenes, perturbations, cellLines,
  method = c("spearman", "pearson", "gsea"), geneSize = 150)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed genes
(e.g. t-statistics)}

\item{geneSize}{Number: top and bottom differentially expressed genes to use
for gene set enrichment (GSE); if \code{method} is not \code{gsea}, this
argument does nothing}
}
\description{
Compare against L1000 datasets
}
\examples{
cellLines <- c("PC3", "VCAP", "A375")
timepoint <- c("96 h")
loadL1000perturbations("L1000", cellLines, timepoint)

# Compare against L1000 using Spearman correlation
compareAgainstL1000(diffExprGenes, perturbations, cellLines,
                    method="spearman")

# Compare against L1000 using Pearson correlation
compareAgainstL1000(diffExprGenes, perturbations, cellLines,
                    method="pearson")

# Compare against L1000 using gene set enrichment analysis (GSEA)
compareAgainstL1000(diffExprGenes, perturbations, cellLines, method="gsea")
}
