% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareAgainstReferencePerMethod}
\alias{compareAgainstReferencePerMethod}
\title{Compare single method}
\usage{
compareAgainstReferencePerMethod(method, diffExprGenes, reference,
  geneSize = 150, cellLines = NULL, cellLineMean = "auto",
  rankPerCellLine = FALSE)
}
\arguments{
\item{method}{Character: comparison methods to run (\code{spearman},
\code{pearson} or \code{gsea}); multiple methods can be selected}

\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
whose names are gene identifiers and respective values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{reference}{Data matrix or \code{perturbationChanges} object (CMap
perturbations; read \code{\link{prepareCMapPerturbations}})}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes for gene set enrichment (only used if \code{method = gsea})}

\item{cellLines}{Integer: number of unique cell lines}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default) the mean score will be
added if more than one cell line is available}

\item{rankPerCellLine}{Boolean: when ranking results, also rank them based on
individual cell lines instead of only focusing on the mean score across
cell lines; if \code{cellLineMean = FALSE}, individual cell line conditions
are always ranked}

\item{rankByAscending}{Boolean: rank values based on their ascending (TRUE)
or descending (FALSE) order?}
}
\value{
Data frame containing the results per method of comparison
}
\description{
Compare single method
}
\keyword{internal}
