% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correlatePerCellLine}
\alias{correlatePerCellLine}
\title{Correlate differential expression scores per cell line}
\usage{
correlatePerCellLine(cellLine, diffExprGenes, perturbations, method)
}
\description{
Correlate differential expression scores per cell line
}
