% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{correlatePerCellLine}
\alias{correlatePerCellLine}
\title{Correlate differential expression scores per cell line}
\usage{
correlatePerCellLine(cellLine, diffExprGenes, perturbations, method)
}
\arguments{
\item{cellLine}{Character: cell line(s)}

\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{l1000perturbations} object: file with L1000 loaded
perturbations (check \code{\link{loadL1000perturbations}})}

\item{method}{Character: correlation method}
}
\value{
Data frame with correlations statistics, p-value and q-value
}
\description{
Correlate differential expression scores per cell line
}
\keyword{internal}
