% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{correlatePerCellLine}
\alias{correlatePerCellLine}
\title{Correlate differential expression scores per cell line}
\usage{
correlatePerCellLine(cellLine, diffExprGenes, perturbations, method)
}
\arguments{
\item{cellLine}{Character: cell line}

\item{diffExprGenes}{Numeric vector: statistic for differentially expressed
genes}

\item{method}{Character: correlation method}
}
\description{
Correlate differential expression scores per cell line
}
