% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.perturbationChanges}
\alias{dim.perturbationChanges}
\title{Dimensions of a \code{perturbationChanges} object}
\usage{
\method{dim}{perturbationChanges}(x)
}
\arguments{
\item{x}{\code{perturbationChanges} object}
}
\value{
Dimensions of a \code{perturbationChanges} object
}
\description{
Dimensions of a \code{perturbationChanges} object
}
