% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimnames.perturbationChanges}
\alias{dimnames.perturbationChanges}
\title{Dimnames of a \code{perturbationChanges} object}
\usage{
\method{dimnames}{perturbationChanges}(x)
}
\arguments{
\item{x}{\code{perturbationChanges} object}
}
\value{
Retrieve dimnames of a \code{perturbationChanges} object
}
\description{
Dimnames of a \code{perturbationChanges} object
}
