% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{downloadENCODEknockdownMetadata}
\alias{downloadENCODEknockdownMetadata}
\title{Download metadata for ENCODE knockdown experiments}
\usage{
downloadENCODEknockdownMetadata(cellLine = NULL, gene = NULL)
}
\arguments{
\item{cellLine}{Character: cell line}

\item{gene}{Character: target gene}
}
\value{
Data frame containing ENCODE knockdown experiment metadata
}
\description{
Download metadata for ENCODE knockdown experiments
}
\examples{
downloadENCODEknockdownMetadata("HepG2", "EIF4G1")
}
