% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{downloadENCODEknockdownMetadata}
\alias{downloadENCODEknockdownMetadata}
\title{Download metadata for ENCODE knockdown experiments}
\usage{
downloadENCODEknockdownMetadata(
  cellLine = NULL,
  gene = NULL,
  file = "ENCODEmetadata.rds"
)
}
\arguments{
\item{cellLine}{Character: cell line}

\item{gene}{Character: target gene}

\item{file}{Character: RDS filepath with metadata (if file doesn't exist, it
will be created)}
}
\value{
Data frame containing ENCODE knockdown experiment metadata
}
\description{
Download metadata for ENCODE knockdown experiments
}
\examples{
downloadENCODEknockdownMetadata("HepG2", "EIF4G1")
}
\seealso{
Other functions related with using ENCODE expression data: 
\code{\link{loadENCODEsamples}()},
\code{\link{performDifferentialExpression}()},
\code{\link{prepareENCODEgeneExpression}()}
}
\concept{functions related with using ENCODE expression data}
