% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadIfNeeded}
\alias{downloadIfNeeded}
\title{Download data if file is not found}
\usage{
downloadIfNeeded(file, link, gz = TRUE)
}
\arguments{
\item{file}{Character: filepath}

\item{link}{Character: link to download file}

\item{gz}{Boolean: is downloaded file compressed?}
}
\value{
Download file if file is not found
}
\description{
Download data if file is not found
}
\keyword{internal}
