% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{findIntersectingCompounds}
\alias{findIntersectingCompounds}
\title{Check for intersecting compounds across specific columns on both datasets}
\usage{
findIntersectingCompounds(data1, data2, keys1 = NULL, keys2 = NULL)
}
\description{
Check for intersecting compounds across specific columns on both datasets
}
\keyword{internal}
