% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{getCMapPerturbationTypes}
\alias{getCMapPerturbationTypes}
\title{Get perturbation types}
\usage{
getCMapPerturbationTypes(control = FALSE)
}
\arguments{
\item{control}{Boolean: return perturbation types used as control?}
}
\value{
Perturbation types and respective codes as used by CMap datasets
}
\description{
Get perturbation types
}
\examples{
getCMapPerturbationTypes()
}
