% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{getL1000Conditions}
\alias{getL1000Conditions}
\title{List available conditions in L1000 datasets}
\usage{
getL1000Conditions(l1000metadataFile)
}
\arguments{
\item{l1000metadataFile}{Character: filepath to L1000 metadata}
}
\value{
List of conditions in L1000 datasets
}
\description{
Downloads metadata if not available
}
\examples{
getL1000Conditions("L1000metadata.txt")
}
