% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{getL1000conditions}
\alias{getL1000conditions}
\title{List available conditions in L1000 datasets}
\usage{
getL1000conditions(metadata, cellLine = NULL, timepoint = NULL,
  dosage = NULL, perturbationType = NULL, control = FALSE)
}
\arguments{
\item{metadata}{Data frame: metadata}

\item{cellLine}{Character: cell line (if \code{NULL}, all values are loaded)}

\item{timepoint}{Character: timepoint (if \code{NULL}, all values are loaded)}

\item{dosage}{Character: dosage (if \code{NULL}, all values are loaded)}

\item{perturbationType}{Character: type of perturbation (if \code{NULL}, all
perturbation types are loaded)}

\item{control}{Boolean: show controls for perturbation types?}
}
\value{
List of conditions in L1000 datasets
}
\description{
Downloads metadata if not available
}
\examples{
data("l1000metadata")
# l1000metadata <- downloadL1000data("l1000metadata.txt", "metadata")
}
