% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchDiffExprLoader}
\alias{launchDiffExprLoader}
\title{Launch differential expression data}
\usage{
launchDiffExprLoader(cellLine = NULL, gene = NULL)
}
\description{
Launch differential expression data
}
