% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchMetadataViewer}
\alias{launchMetadataViewer}
\title{View metadata}
\usage{
launchMetadataViewer(...)
}
\arguments{
\item{...}{Objects}
}
\value{
Launches metadata viewer
}
\description{
View metadata
}
