% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInterface.R
\name{launchResultPlotter}
\alias{launchResultPlotter}
\title{View and plot results via a visual interface}
\usage{
launchResultPlotter(...)
}
\arguments{
\item{...}{Objects}
}
\value{
Launches result viewer and plotter (returns \code{NULL})
}
\description{
View and plot results via a visual interface
}
\seealso{
Other visual interface functions: 
\code{\link{cTRAP}()},
\code{\link{launchCMapDataLoader}()},
\code{\link{launchDiffExprLoader}()},
\code{\link{launchDrugSetEnrichmentAnalyser}()},
\code{\link{launchMetadataViewer}()}
}
\concept{visual interface functions}
