% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{loadCMapPerturbations}
\alias{loadCMapPerturbations}
\title{Load CMap perturbation data}
\usage{
loadCMapPerturbations(metadata, zscores, geneInfo, compoundInfo = NULL)
}
\arguments{
\item{metadata}{Data frame (CMap metadata) or character (respective filepath
to load data from file)}

\item{zscores}{Data frame (GCTX z-scores) or character (respective filepath
to load data from file)}

\item{geneInfo}{Data frame (CMap gene info) or character (respective
filepath to load data from file)}

\item{compoundInfo}{Data frame (CMap compound info) or character (respective
filepath to load data from file)}
}
\value{
Perturbation data from CMap as a data table
}
\description{
Load CMap perturbation data
}
\examples{
if (interactive()) {
  metadata <- loadCMapData("cmapMetadata.txt", "metadata")
  metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
  loadCMapPerturbations(metadata, "cmapZscores.gctx", "cmapGeneInfo.txt")
}
}
