% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{loadCMapZscores}
\alias{loadCMapZscores}
\title{Load matrix of CMap perturbation's differential expression z-scores}
\usage{
loadCMapZscores(data, perturbationChanges = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{\code{perturbationChanges} object}

\item{perturbationChanges}{Boolean: convert to \code{perturbationChanges}
object?}

\item{verbose}{Boolean: print messages?}
}
\value{
Matrix containing CMap perturbation z-scores (genes as rows,
  perturbations as columns)
}
\description{
Load matrix of CMap perturbation's differential expression z-scores
}
\examples{
\donttest{
metadata <- loadCMapData("cmapMetadata.txt", "metadata")
metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
perts <- prepareCMapPerturbations(metadata, "cmapZscores.gctx",
                                  "cmapGeneInfo.txt")
zscores <- loadCMapZscores(perts[ , 1:10])
}
}
