% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{loadCMapZscores}
\alias{loadCMapZscores}
\title{Load matrix of CMap zscores}
\usage{
loadCMapZscores(data, verbose = TRUE)
}
\arguments{
\item{data}{cmapPerturbations object}

\item{verbose}{Boolean: print messages?}
}
\value{
Matrix containing CMap perturbation z-scores (genes as rows,
  perturbations as columns)
}
\description{
Load matrix of CMap zscores
}
\examples{
#' if (interactive()) {
  metadata <- loadCMapData("cmapMetadata.txt", "metadata")
  metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
  perts <- prepareCMapPerturbations(metadata, "cmapZscores.gctx",
                                    "cmapGeneInfo.txt")
  zscores <- loadCMapZscores(perts[ , 1:10])
}
}
