% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{loadCTRPgeneExpression}
\alias{loadCTRPgeneExpression}
\alias{loadCTRPdrugSensitivity}
\alias{loadCTRPcompoundInfo}
\alias{loadNCI60geneExpression}
\alias{loadGDSCfile}
\alias{loadGDSCcellLineInfo}
\alias{loadGDSCcompoundInfo}
\alias{loadGDSCgeneExpression}
\alias{loadGDSCdrugSensitivity}
\title{Load CTRP data}
\usage{
loadCTRPgeneExpression(geneExpressionFile = "CTRP/geneExpr.txt",
  geneInfoFile = "CTRP/geneInfo.txt",
  cellLineInfoFile = "CTRP/cellLineInfo.txt")


  loadCTRPdrugSensitivity(drugSensitivityFile = "CTRP/drugSensitivity.txt",
  experimentFile = "CTRP/experimentInfo.txt",
  compoundFile = "CTRP/compoundInfo.txt")

loadCTRPcompoundInfo(compoundFile = "CTRP/compoundInfo.txt")

loadNCI60geneExpression(file = "NCI60/geneExpr.xls",
  cellLineInfoFile = "cellLineInfo.xls")

loadGDSCfile(file, filename, type, ...)

loadGDSCcellLineInfo(file = "GDSC/cellLineInfo.xlsx")

loadGDSCcompoundInfo(file = "GDSC/compoundInfo.xlsx")

loadGDSCgeneExpression(file = "GDSC/geneExpr.txt")

loadGDSCdrugSensitivity(file = "GDSC/drugs.xlsx")
}
\arguments{
\item{geneExpressionFile}{Character: path to file with gene expression}

\item{geneInfoFile}{Character: path to file with gene information}

\item{cellLineInfoFile}{Character: path to file with cell line information}

\item{drugSensitivityFile}{Character: path to file with drug sensitivity}

\item{experimentFile}{Character: path to file with experiment information}

\item{compoundFile}{Character: path to file with compound information}

\item{file}{Character: file path}
}
\value{
Data frame
}
\description{
If given paths direct to non-existing files, those files will be downloaded
}
\keyword{internal}
