% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{loadDrugDescriptors}
\alias{loadDrugDescriptors}
\title{Load table with drug descriptors}
\usage{
loadDrugDescriptors()
}
\value{
Data table with drug descriptors (first column, named )
}
\description{
Load table with drug descriptors
}
\examples{
loadDrugDescriptors()
}
