% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.R
\name{loadENCODEsample}
\alias{loadENCODEsample}
\title{Load ENCODE sample}
\usage{
loadENCODEsample(metadata, replicate, control = FALSE, path = ".")
}
\arguments{
\item{metadata}{Data frame: ENCODE metadata}

\item{replicate}{Number: replicate}

\item{control}{Boolean: load control experiment?}

\item{path}{Character: path where to download files}
}
\value{
Data table with ENCODE sample data
}
\description{
Load ENCODE sample
}
\keyword{internal}
