% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{loadL1000perturbations}
\alias{loadL1000perturbations}
\title{Load L1000 perturbation data}
\usage{
loadL1000perturbations(metadata, zscores, geneInfo,
  sanitizeCompoundNames = FALSE)
}
\arguments{
\item{metadata}{Data frame: L1000 Metadata}

\item{zscores}{Data frame: GCTX z-scores}

\item{geneInfo}{Data frame: L1000 gene info}

\item{sanitizeCompoundNames}{Boolean: replace identifiers with compound names}
}
\value{
Perturbation data from L1000 as a data table
}
\description{
Load L1000 perturbation data
}
\examples{
if (interactive()) {
  metadata <- downloadL1000metadata("l1000metadata.txt")
  metadata <- filterL1000metadata(metadata, cellLine="HepG2")
  zscores  <- downloadL1000zscores("l1000zscores.gctx", metadata$sig_id)
  geneInfo <- downloadL1000geneInfo("l1000geneInfo.txt")
  loadL1000perturbations(metadata, zscores, geneInfo)
}
}
