% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{matchStatsWithDrugSetsID}
\alias{matchStatsWithDrugSetsID}
\title{Match identifiers between stats and drug sets}
\usage{
matchStatsWithDrugSetsID(sets, stats, col = NULL)
}
\description{
Match identifiers between stats and drug sets
}
\keyword{internal}
