% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{performDifferentialExpression}
\alias{performDifferentialExpression}
\title{Perform differential gene expression based on ENCODE data}
\usage{
performDifferentialExpression(counts)
}
\arguments{
\item{counts}{Data frame: gene expression}
}
\value{
Data frame with differential gene expression results between
knockdown and control
}
\description{
Perform differential gene expression based on ENCODE data
}
\examples{
gene <- "EIF4G1"
cellLine <- "HepG2"

ENCODEmetadata <- downloadENCODEknockdownMetadata(cellLine, gene)
counts <- loadENCODEgeneExpression(ENCODEmetadata)

# Remove low coverage (at least 10 counts shared across two samples)
minReads   <- 10
minSamples <- 2
filter <- rowSums(counts[ , 3:6] >= minReads) >= minSamples
counts <- counts[filter, ]

# Perform differential gene expression analysis
diffExpr <- performDifferentialExpression(counts)
}
