% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{performDifferentialExpression}
\alias{performDifferentialExpression}
\title{Perform differential gene expression}
\usage{
performDifferentialExpression(counts, geneAnnot)
}
\description{
Perform differential gene expression
}
