% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{performDifferentialExpression}
\alias{performDifferentialExpression}
\title{Perform differential gene expression based on ENCODE data}
\usage{
performDifferentialExpression(counts, geneAnnot)
}
\arguments{
\item{counts}{Data frame: gene expression}

\item{geneAnnot}{Data frame: gene annotation}
}
\value{
Data frame with differential gene expression results between
knockdown and control
}
\description{
Perform differential gene expression based on ENCODE data
}
