% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{performGSAperCellLine}
\alias{performGSAperCellLine}
\title{Perform gene set enrichment (GSA) per cell line}
\usage{
performGSAperCellLine(cellLine, countTable, perturbations, gsc)
}
\description{
Perform gene set enrichment (GSA) per cell line
}
