% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1000.R
\name{performGSAperCellLine}
\alias{performGSAperCellLine}
\title{Perform gene set enrichment (GSA) per cell line}
\usage{
performGSAperCellLine(cellLine, countTable, perturbations, gsc)
}
\value{
Data frame containing gene set enrichment analysis (GSEA) results per
cell line
}
\description{
Perform gene set enrichment (GSA) per cell line
}
