% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{performGSEAagainstReference}
\alias{performGSEAagainstReference}
\title{Perform gene set enrichment (GSA) against data columns}
\usage{
performGSEAagainstReference(reference, geneset)
}
\value{
Data frame containing gene set enrichment analysis (GSEA) results per
data column
}
\description{
Perform gene set enrichment (GSA) against data columns
}
\keyword{internal}
