% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cmapComparison}
\alias{plot.cmapComparison}
\title{Plot CMap data comparison}
\usage{
\method{plot}{cmapComparison}(x, method = c("spearman", "pearson",
  "gsea"), n = c(3, 3), showMetadata = TRUE, alpha = 0.3)
}
\arguments{
\item{x}{\code{cmapComparison} object}

\item{method}{Character: method to plot results (\code{spearman},
\code{pearson} or \code{gsea})}

\item{n}{Numeric: number of top and bottom genes to label (if a vector of two
numbers is given, the first and second numbers will be used as the number
of top and bottom genes to label, respectively)}

\item{showMetadata}{Boolean: show available metadata information instead of
perturbation identifiers?}

\item{alpha}{Numeric: transparency}
}
\value{
Plot illustrating the comparison with CMap data
}
\description{
Plot CMap data comparison
}
\examples{
data("cmapPerturbationsKD")

# Compare against CMap using Spearman's correlation, Pearson's correlation
# and gene set enrichment analysis (GSEA) with the top and bottom 150 genes
# as gene sets
compareKD <- compareAgainstCMap(diffExprStat, cmapPerturbationsKD)

plot(compareKD, "spearman", c(7, 3))
plot(compareKD, "pearson")
plot(compareKD, "gsea")
}
