% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.referenceComparison}
\alias{plot.referenceComparison}
\title{Plot data comparison}
\usage{
\method{plot}{referenceComparison}(
  x,
  method = c("spearman", "pearson", "gsea", "rankProduct"),
  n = c(3, 3),
  showMetadata = TRUE,
  plotNonRankedPerturbations = FALSE,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{x}{\code{referenceComparison} object: obtained after running
\code{\link{rankSimilarPerturbations}} or
\code{\link{predictTargetingDrugs}}}

\item{method}{Character: method to plot results (\code{spearman},
\code{pearson}, \code{gsea} or \code{rankProduct})}

\item{n}{Numeric: number of top and bottom genes to label (if a vector of two
numbers is given, the first and second numbers will be used as the number
of top and bottom genes to label, respectively)}

\item{showMetadata}{Boolean: show available metadata information instead of
identifiers (if available)?}

\item{plotNonRankedPerturbations}{Boolean: plot non-ranked data in grey?}

\item{alpha}{Numeric: transparency}

\item{...}{Extra arguments currently not used}
}
\value{
Plot illustrating the reference comparison
}
\description{
Plot data comparison
}
\examples{
# Example of a differential expression profile
data("diffExprStat")

\dontrun{
# Download and load CMap perturbations to compare with
cellLine <- "HepG2"
cmapMetadataKD <- filterCMapMetadata(
  "cmapMetadata.txt", cellLine=cellLine,
  perturbationType="Consensus signature from shRNAs targeting the same gene")

cmapPerturbationsKD <- prepareCMapPerturbations(
  cmapMetadataKD, "cmapZscores.gctx", "cmapGeneInfo.txt", loadZscores=TRUE)
}

# Rank similar CMap perturbations
compareKD <- rankSimilarPerturbations(diffExprStat, cmapPerturbationsKD)

plot(compareKD, "spearman", c(7, 3))
plot(compareKD, "pearson")
plot(compareKD, "gsea")
}
\seealso{
Other functions related with the ranking of CMap perturbations: 
\code{\link{[.perturbationChanges}()},
\code{\link{as.table.similarPerturbations}()},
\code{\link{dim.perturbationChanges}()},
\code{\link{dimnames.perturbationChanges}()},
\code{\link{filterCMapMetadata}()},
\code{\link{getCMapConditions}()},
\code{\link{getCMapPerturbationTypes}()},
\code{\link{loadCMapData}()},
\code{\link{loadCMapZscores}()},
\code{\link{parseCMapID}()},
\code{\link{plot.perturbationChanges}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{prepareCMapPerturbations}()},
\code{\link{print.similarPerturbations}()},
\code{\link{rankSimilarPerturbations}()}

Other functions related with the prediction of targeting drugs: 
\code{\link{listExpressionDrugSensitivityAssociation}()},
\code{\link{loadExpressionDrugSensitivityAssociation}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{predictTargetingDrugs}()}
}
\concept{functions related with the prediction of targeting drugs}
\concept{functions related with the ranking of CMap perturbations}
