% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{plotDrugSetEnrichment}
\alias{plotDrugSetEnrichment}
\title{Plot drug set enrichment}
\usage{
plotDrugSetEnrichment(sets, stats, col = "rankProduct_rank")
}
\arguments{
\item{sets}{Named list of characters: named sets containing compound
identifiers (obtain drug sets by running \code{prepareDrugSets()})}

\item{stats}{Named vector of characters or either a
\code{similarPerturbations} or a \code{targetingDrugs} object (obtained
after running \code{\link{rankSimilarPerturbations}} or
\code{\link{predictTargetingDrugs}}, respectively)}

\item{col}{Character: name of the column to use for statistics (only required
if class of \code{stats} is either \code{similarPerturbations} or
\code{targetingDrugs})}
}
\value{
List of GSEA plots per drug set
}
\description{
Plot drug set enrichment
}
\seealso{
Other functions for drug set enrichment analysis: \code{\link{analyseDrugSetEnrichment}},
  \code{\link{prepareDrugSets}}
}
\concept{functions for drug set enrichment analysis}
