% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot gene set enrichment analysis (GSEA)}
\usage{
plotGSEA(stats, topGenes = NULL, bottomGenes = NULL,
  title = "GSEA plot", gseaParam = 1)
}
\arguments{
\item{stats}{Named numeric vector: statistics}

\item{topGenes}{Named list of characters: up-regulated genes}

\item{bottomGenes}{Named list of characters: down-regulated genes}

\item{title}{Character: title}

\item{gseaParam}{Numeric: GSEA-like parameter}
}
\value{
Grid of plots illustrating a GSEA plot
}
\description{
Plot gene set enrichment analysis (GSEA)
}
\keyword{internal}
