% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot gene set enrichment analysis (GSEA)}
\usage{
plotGSEA(perturbation, pathways, genes = c("both", "top", "bottom"),
  title = "GSEA plot", titleSize = 14, axisTitleSize = 12,
  axisTextSize = 10, gseaParam = 1)
}
\arguments{
\item{perturbation}{List of named numeric vectors containing the differential
expression profile score per gene for a perturbation; each perturbation of
the list will be rendered with a different colour}

\item{pathways}{Named list of characters: up-regulated (named \code{top}) and
down-regulated (\code{bottom}) genes}

\item{title}{Character: title}

\item{titleSize}{Integer: title size}

\item{gseaParam}{Numeric: GSEA-like parameter}
}
\value{
Grid of plots illustrating a GSEA plot
}
\description{
Plot gene set enrichment analysis (GSEA)
}
\keyword{internal}
