% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot gene set enrichment analysis (GSEA)}
\usage{
plotGSEA(perturbation, pathways, genes = c("both", "top", "bottom"),
  titleSize = 14, axisTitleSize = 12, axisTextSize = 10,
  enrichmentPlotPointSize = 0.1, gseaParam = 1)
}
\arguments{
\item{perturbation}{List of named numeric vectors containing the differential
expression profile score per gene for a perturbation; each perturbation of
the list will be rendered with a different colour}

\item{pathways}{List of gene sets to check.}

\item{titleSize}{Integer: plot title size}

\item{gseaParam}{Numeric: GSEA-like parameter}

\item{title}{Character: plot title}
}
\value{
Grid of plots illustrating a GSEA plot
}
\description{
Plot gene set enrichment analysis (GSEA)
}
\keyword{internal}
