% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot gene set enrichment analysis (GSEA)}
\usage{
plotGSEA(pathways, stats, title = "GSEA plot")
}
\arguments{
\item{pathways}{List of gene sets to check.}

\item{stats}{Named vector of gene-level stats. Names should be the same as in 'pathways'}

\item{title}{Character: plot title}
}
\value{
Grid of plots illustrating a GSEA plot
}
\description{
Plot gene set enrichment analysis (GSEA)
}
