% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotL1000comparison}
\alias{plotL1000comparison}
\title{Plot L1000 data comparison}
\usage{
plotL1000comparison(object, perturbationID, topGenes = TRUE)
}
\arguments{
\item{object}{\code{L1000 comparison} object}

\item{perturbationID}{Character: perturbation identifier}

\item{topGenes}{Boolean: plot top (\code{topGenes = TRUE}) or bottom genes
(\code{topGenes = FALSE}) in case of plotting gene set enrichment analysis
(GSEA)}
}
\value{
Plot illustrating the comparison with L1000 data
}
\description{
Plot L1000 data comparison
}
\examples{
compareKnockdown <- loadInternalData("compareKnockdown")
EIF4G1knockdown <- grep("EIF4G1", compareKnockdown$gsea_ordered$genes,
                        value=TRUE)
plotL1000comparison(compareKnockdown$spearman, EIF4G1knockdown)
plotL1000comparison(compareKnockdown$pearson, EIF4G1knockdown)
plotL1000comparison(compareKnockdown$gsea, EIF4G1knockdown)
}
