% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotL1000comparison}
\alias{plotL1000comparison}
\alias{plot.l1000comparison}
\title{Plot L1000 data comparison}
\usage{
plotL1000comparison(x, perturbation = NULL, perturbationID = NULL,
  genes = c("both", "top", "bottom"))

\method{plot}{l1000comparison}(x, perturbation = NULL,
  perturbationID = NULL, genes = c("both", "top", "bottom"))
}
\arguments{
\item{x}{\code{l1000comparison} object}

\item{perturbationID}{Character: perturbation identifier}

\item{genes}{Character: when plotting gene set enrichment analysis (GSEA),
plot top genes (\code{genes = "top"}), bottom genes
(\code{genes = "bottom"}) or both (\code{genes = "both"})}
}
\value{
Plot illustrating the comparison with L1000 data
}
\description{
Plot L1000 data comparison
}
\examples{
data("l1000perturbationsKnockdown")
cellLine <- "HepG2"
compareKnockdown <- list()

# Compare against L1000 using Spearman correlation
compareKnockdown$spearman <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="spearman")

# Compare against L1000 using Pearson correlation
compareKnockdown$pearson <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="pearson")

# Compare against L1000 using gene set enrichment analysis (GSEA) with the
# top and bottom 150 genes as gene sets
compareKnockdown$gsea <- compareAgainstL1000(
    diffExprStat, l1000perturbationsKnockdown, cellLine, method="gsea",
    geneSize=150)

EIF4G1knockdown <- grep("EIF4G1", compareKnockdown$gsea$genes, value=TRUE)
plotL1000comparison(compareKnockdown$spearman, EIF4G1knockdown)
plotL1000comparison(compareKnockdown$pearson, EIF4G1knockdown)
plotL1000comparison(compareKnockdown$gsea, EIF4G1knockdown)
}
