% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMetricDistribution}
\alias{plotMetricDistribution}
\title{Plot metric distribution}
\usage{
plotMetricDistribution(statsOrd, breaks = 50, axisTitleSize = 12,
  axisTextSize = 10)
}
\value{
Metric distribution plot
}
\description{
Plot metric distribution
}
\keyword{internal}
