% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSingleCorr}
\alias{plotSingleCorr}
\title{Render scatter plot to show a single relationship}
\usage{
plotSingleCorr(perturbation, ylabel, diffExprGenes, title = NULL)
}
\arguments{
\item{perturbation}{List of named numeric vectors containing the differential
expression profile score per gene for a perturbation; each perturbation of
the list will be rendered with a different colour}

\item{ylabel}{Character: Y axis label}

\item{diffExprGenes}{Named numeric vector}

\item{title}{Character: plot title (if \code{NULL}, the default title depends
on the context; ignored when plotting multiple perturbations)}
}
\value{
Scatter plot
}
\description{
Render scatter plot to show a single relationship
}
\keyword{internal}
