% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTargetingDrugsVSsimilarPerturbations}
\alias{plotTargetingDrugsVSsimilarPerturbations}
\title{Plot similar perturbations against predicted targeting drugs}
\usage{
plotTargetingDrugsVSsimilarPerturbations(
  targetingDrugs,
  similarPerturbations,
  column,
  labelBy = "pert_iname",
  quantileThreshold = 0.25,
  showAllScores = FALSE
)
}
\arguments{
\item{targetingDrugs}{\code{targetingDrugs} object}

\item{similarPerturbations}{\code{similarPerturbations} object}

\item{column}{Character: column to plot (must be available in both databases)}

\item{labelBy}{Character: column in \code{similarPerturbations}, its metadata
or compound information to be used for labelling}

\item{quantileThreshold}{Numeric: quantile to use for highlight values within
[0, 1]}

\item{showAllScores}{Boolean: show all scores? If \code{FALSE}, only the best
score per compound will be plotted}
}
\value{
\code{ggplot2} plot
}
\description{
Plot similar perturbations against predicted targeting drugs
}
\seealso{
Other functions related with the ranking of CMap perturbations: 
\code{\link{as.table.referenceComparison}()},
\code{\link{filterCMapMetadata}()},
\code{\link{getCMapConditions}()},
\code{\link{getCMapPerturbationTypes}()},
\code{\link{loadCMapData}()},
\code{\link{loadCMapZscores}()},
\code{\link{parseCMapID}()},
\code{\link{plot.perturbationChanges}()},
\code{\link{plot.referenceComparison}()},
\code{\link{prepareCMapPerturbations}()},
\code{\link{print.similarPerturbations}()},
\code{\link{rankSimilarPerturbations}()}

Other functions related with the prediction of targeting drugs: 
\code{\link{as.table.referenceComparison}()},
\code{\link{listExpressionDrugSensitivityAssociation}()},
\code{\link{loadExpressionDrugSensitivityAssociation}()},
\code{\link{plot.referenceComparison}()},
\code{\link{predictTargetingDrugs}()}
}
\concept{functions related with the prediction of targeting drugs}
\concept{functions related with the ranking of CMap perturbations}
