% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivity.R
\name{predictTargetingDrugs}
\alias{predictTargetingDrugs}
\title{Predict targeting drugs}
\usage{
predictTargetingDrugs(
  input,
  expressionDrugSensitivityCor,
  method = c("spearman", "pearson", "gsea"),
  geneSize = 150,
  isDrugActivityDirectlyProportionalToSensitivity = NULL
)
}
\arguments{
\item{input}{\code{Named numeric vector} of differentially expressed genes
whose names are gene identifiers and respective values are a statistic that
represents significance and magnitude of differentially expressed genes
(e.g. t-statistics); or \code{character} of gene symbols composing a gene
set that is tested for enrichment in reference data (only used if
\code{method} includes \code{gsea})}

\item{expressionDrugSensitivityCor}{Matrix: correlation matrix of gene
expression (rows) and drug sensitivity (columns) across cell lines.
Pre-prepared gene expression and drug sensitivity associations are
available to download using
\code{\link{loadExpressionDrugSensitivityAssociation}()}.}

\item{method}{Character: one or more methods to compare data
(\code{spearman}, \code{pearson} or \code{gsea})}

\item{geneSize}{Numeric: number of top up-/down-regulated genes to use as
gene sets to test for enrichment in reference data; if a 2-length numeric
vector, the first index is the number of top up-regulated genes and the
second index is the number of down-regulated genes used to create gene
sets; only used if \code{method} includes \code{gsea} and if \code{input}
is not a gene set}

\item{isDrugActivityDirectlyProportionalToSensitivity}{Boolean: are the
values used for drug activity directly proportional to drug sensitivity?
See details.}
}
\value{
Data table with correlation or GSEA results comparing differential
  expression values against gene expression and drug sensitivity associations
}
\description{
Identify compounds that may target the phenotype associated with a
user-provided differential expression profile by comparing such against a
correlation matrix of gene expression and drug sensitivity.
}
\details{
If \code{isDrugActivityDirectlyProportionalToSensitivity = NULL}, the
  attribute \code{isDrugMetricDirectlyProportionalToSensitivity} of
  \code{expressionDrugSensitivityCor} is used (see
  \code{\link{loadExpressionDrugSensitivityAssociation}()}).
}
\section{GSEA score}{

Weighted connectivity scores (WTCS) are calculated when \code{method
  = "gsea"} (\url{https://clue.io/connectopedia/cmap_algorithms}).
}

\examples{
# Example of a differential expression profile
data("diffExprStat")

# Load expression and drug sensitivity association derived from GDSC data
gdsc <- loadExpressionDrugSensitivityAssociation("GDSC 7")

# Predict targeting drugs on a differential expression profile
predictTargetingDrugs(diffExprStat, gdsc)
}
\seealso{
Other functions related with the prediction of targeting drugs: 
\code{\link{as.table.targetingDrugs}()},
\code{\link{listExpressionDrugSensitivityAssociation}()},
\code{\link{loadExpressionDrugSensitivityAssociation}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()}
}
\concept{functions related with the prediction of targeting drugs}
