% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{prepareCMapPerturbations}
\alias{prepareCMapPerturbations}
\title{Prepare CMap perturbation data}
\usage{
prepareCMapPerturbations(
  metadata,
  zscores,
  geneInfo,
  compoundInfo = NULL,
  loadZscores = FALSE
)
}
\arguments{
\item{metadata}{Data frame (CMap metadata) or character (respective filepath
to load data from file)}

\item{zscores}{Data frame (GCTX z-scores) or character (respective filepath
to load data from file)}

\item{geneInfo}{Data frame (CMap gene info) or character (respective
filepath to load data from file)}

\item{compoundInfo}{Data frame (CMap compound info) or character (respective
filepath to load data from file)}

\item{loadZscores}{Boolean: load perturbation z-scores? Not recommended in
memory-constrained systems}
}
\value{
CMap perturbation data attributes and filename
}
\description{
Prepare CMap perturbation data
}
\examples{
\dontrun{
metadata <- loadCMapData("cmapMetadata.txt", "metadata")
metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
prepareCMapPerturbations(metadata, "cmapZscores.gctx", "cmapGeneInfo.txt")
}
}
\seealso{
Other functions related with the ranking of CMap perturbations: 
\code{\link{as.table.similarPerturbations}()},
\code{\link{filterCMapMetadata}()},
\code{\link{getCMapConditions}()},
\code{\link{getCMapPerturbationTypes}()},
\code{\link{loadCMapData}()},
\code{\link{loadCMapZscores}()},
\code{\link{parseCMapID}()},
\code{\link{plot.perturbationChanges}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{print.similarPerturbations}()},
\code{\link{rankSimilarPerturbations}()}
}
\concept{functions related with the ranking of CMap perturbations}
