% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{prepareGSEApathways}
\alias{prepareGSEApathways}
\title{Prepare GSEA pathways}
\usage{
prepareGSEApathways(diffExprGenes, geneSize)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes to use for gene set enrichment (GSE) (only used if
\code{method = gsea})}
}
\value{
List of top and bottom differentially expressed genes
}
\description{
Prepare GSEA pathways
}
\keyword{internal}
