% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{prepareSetsCompoundInfo}
\alias{prepareSetsCompoundInfo}
\title{Get drug sets' compound info}
\usage{
prepareSetsCompoundInfo(sets)
}
\arguments{
\item{sets}{Named list of characters: named sets containing compound
identifiers (obtain drug sets by running \code{prepareDrugSets()})}
}
\value{
List containing drug sets' compound info
}
\description{
Get drug sets' compound info
}
\keyword{internal}
