% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSetEnrichment.R
\name{prepareStatsCompoundInfo}
\alias{prepareStatsCompoundInfo}
\title{Prepare stats' compound information}
\usage{
prepareStatsCompoundInfo(stats)
}
\arguments{
\item{stats}{Named numeric vector or either a \code{similarPerturbations} or
a \code{targetingDrugs} object (obtained after running
\code{\link{rankSimilarPerturbations}} or
\code{\link{predictTargetingDrugs}}, respectively)}
}
\value{
List containing stats' compound info
}
\description{
Prepare stats' compound information
}
\keyword{internal}
