% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.similarPerturbations}
\alias{print.similarPerturbations}
\title{Print a \code{similarPerturbations} object}
\usage{
\method{print}{similarPerturbations}(x, perturbation = NULL, ...)
}
\arguments{
\item{x}{\code{similarPerturbations} object}

\item{perturbation}{Character (perturbation identifier) or numeric
(perturbation index)}

\item{...}{Extra parameters passed to \code{print}}
}
\value{
Information on \code{perturbationChanges} object or on specific
  perturbations (if \code{perturbation} is set)
}
\description{
Print a \code{similarPerturbations} object
}
