% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{print.similarPerturbations}
\alias{print.similarPerturbations}
\title{Print a \code{similarPerturbations} object}
\usage{
\method{print}{similarPerturbations}(x, perturbation = NULL, ...)
}
\arguments{
\item{x}{\code{similarPerturbations} object}

\item{perturbation}{Character (perturbation identifier) or numeric
(perturbation index)}

\item{...}{Extra parameters passed to \code{print}}
}
\value{
Information on \code{perturbationChanges} object or on specific
  perturbations
}
\description{
Print a \code{similarPerturbations} object
}
\seealso{
Other functions related with the ranking of CMap perturbations: 
\code{\link{as.table.similarPerturbations}()},
\code{\link{filterCMapMetadata}()},
\code{\link{getCMapConditions}()},
\code{\link{getCMapPerturbationTypes}()},
\code{\link{loadCMapData}()},
\code{\link{loadCMapZscores}()},
\code{\link{parseCMapID}()},
\code{\link{plot.perturbationChanges}()},
\code{\link{plot.referenceComparison}()},
\code{\link{plotTargetingDrugsVSsimilarPerturbations}()},
\code{\link{prepareCMapPerturbations}()},
\code{\link{rankSimilarPerturbations}()}
}
\concept{functions related with the ranking of CMap perturbations}
