% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{rankAgainstReference}
\alias{rankAgainstReference}
\title{Compare multiple methods and rank against reference accordingly}
\usage{
rankAgainstReference(
  input,
  reference,
  method = c("spearman", "pearson", "gsea"),
  geneSize = 150,
  cellLines = NULL,
  cellLineMean = "auto",
  rankByAscending = TRUE,
  rankPerCellLine = FALSE,
  threads = 1,
  verbose = FALSE
)
}
\arguments{
\item{input}{\code{Named numeric vector} of differentially expressed genes
whose names are gene identifiers and respective values are a statistic that
represents significance and magnitude of differentially expressed genes
(e.g. t-statistics); or \code{character} of gene symbols composing a gene
set that is tested for enrichment in reference data (only used if
\code{method} includes \code{gsea})}

\item{reference}{Data matrix or \code{character} object with file path to
CMap perturbations (see \code{\link{prepareCMapPerturbations}()}) or gene
expression and drug sensitivity association (see
\code{\link{loadExpressionDrugSensitivityAssociation}()})}

\item{method}{Character: one or more methods to compare data
(\code{spearman}, \code{pearson} or \code{gsea})}

\item{geneSize}{Numeric: number of top up-/down-regulated genes to use as
gene sets to test for enrichment in reference data; if a 2-length numeric
vector, the first index is the number of top up-regulated genes and the
second index is the number of down-regulated genes used to create gene
sets; only used if \code{method} includes \code{gsea} and if \code{input}
is not a gene set}

\item{cellLines}{Integer: number of unique cell lines}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default), the mean score will be
added when data for more than one cell line is available.}

\item{rankByAscending}{Boolean: rank values based on their ascending
(\code{TRUE}) or descending (\code{FALSE}) order?}

\item{rankPerCellLine}{Boolean: rank results based on both individual cell
lines and mean scores across cell lines (\code{TRUE}) or based on mean
scores alone (\code{FALSE})? If \code{cellLineMean = FALSE}, individual
cell line conditions are always ranked.}

\item{threads}{Integer: number of parallel threads}

\item{verbose}{Boolean: print additional details?}
}
\value{
List of data frame containing the results per methods of comparison
}
\description{
Compare multiple methods and rank against reference accordingly
}
\keyword{internal}
