% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap.R
\name{rankSimilarPerturbations}
\alias{rankSimilarPerturbations}
\title{Compare differential expression results against CMap perturbations}
\usage{
rankSimilarPerturbations(diffExprGenes, perturbations,
  method = c("spearman", "pearson", "gsea"), geneSize = 150,
  cellLineMean = "auto", rankCellLinePerturbations = FALSE)
}
\arguments{
\item{diffExprGenes}{Numeric: named vector of differentially expressed genes
where the name of the vector are gene names and the values are a statistic
that represents significance and magnitude of differentially expressed
genes (e.g. t-statistics)}

\item{perturbations}{\code{perturbationChanges} object: CMap perturbations
(check \code{\link{prepareCMapPerturbations}})}

\item{method}{Character: comparison method (\code{spearman}, \code{pearson}
or \code{gsea}; multiple methods may be selected at once)}

\item{geneSize}{Number: top and bottom number of differentially expressed
genes to use for gene set enrichment (GSE) (only used if
\code{method = gsea})}

\item{cellLineMean}{Boolean: add a column with the mean score across cell
lines? If \code{cellLineMean = "auto"} (default) the mean score will be
added if more than one cell line is available}

\item{rankCellLinePerturbations}{Boolean: when ranking results, also rank
perturbations regarding individual cell lines instead of the mean
perturbation score alone; if \code{cellLineMean = FALSE}, individual cell
line perturbations are always ranked}
}
\value{
Data table with correlation or GSEA results comparing differential
gene expression values with those associated with CMap perturbations
}
\description{
Weighted connectivity scores (WTCS) are calculated when
\code{method = "gsea"}. For more information on WTCS, read
\url{https://clue.io/connectopedia/cmap_algorithms}.
}
\examples{
data("cmapPerturbationsCompounds")
perturbations <- cmapPerturbationsCompounds
data("diffExprStat")

# Compare differential expression results against CMap perturbations
rankSimilarPerturbations(diffExprStat, perturbations)

# Compare using only Spearman's correlation
rankSimilarPerturbations(diffExprStat, perturbations, method="spearman")
}
