% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.cmapPerturbations}
\alias{[.cmapPerturbations}
\title{Subset a \code{cmapPerturbations} object}
\usage{
\method{[}{cmapPerturbations}(x, i, j, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{cmapPerturbations} object}

\item{i, j}{Character or numeric indexes specifying elements to extract}

\item{drop}{Boolean: coerce result to the lowest possible dimension?}

\item{...}{Extra parameters passed to \code{`[`}}
}
\value{
\code{cmapPerturbations} object with subset data
}
\description{
Subset a \code{cmapPerturbations} object
}
