% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.perturbationChanges}
\alias{[.perturbationChanges}
\title{Subset a \code{perturbationChanges} object}
\usage{
\method{[}{perturbationChanges}(x, i, j, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{perturbationChanges} object}

\item{i, j}{Character or numeric indexes specifying elements to extract}

\item{drop}{Boolean: coerce result to the lowest possible dimension?}

\item{...}{Extra parameters passed to \code{`[`}}
}
\value{
\code{perturbationChanges} object with subset data
}
\description{
Subset a \code{perturbationChanges} object
}
